Add E_STRICT warning code for cases of bad practices
Summary:
E_STRICT is used to indicate a bad practice that one shouldn't
do, generally for things that can cause bad performance but
are allowed.

Adding this for the case that we don't have a default timezone
set since this causes /etc/localtime to be checked each time.

With PHP 5.3, this became a warning.

Test Plan:
build
fast_tests

  <?php
  error_reporting(E_ALL ^ E_STRICT);
  var_dump(date('r'));

DiffCamp Revision: 118748
Reviewed By: andrewparoski
Commenters: hzhao
CC: hphp-diffs@lists, hzhao, macvicar, andrewparoski
Revert Plan:
Ok