[Fix]Make $this point to the root object when dealing with redeclared classes
Summary:
$this was pointing to the this belonging to the object owning the current
method.
It should be the root. Since $this is taken from the frame injection, I changed
it to
also store the root object. I added a virtual method getRoot() to do this.

There's also a fix for dynCreate where it should inc the ref count
of the object before calling the constructor to prevent premature
deallocation.

Test Plan:
TestCodeRun

DiffCamp Revision: 117651
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK