ThreadLocal improvements
Summary:
This revision changes ThreadLocal<T> to be a POD (plain old data)
type, and changes the associated macros to put the "__thread"
keyword in front of ThreadLocal declarations and definitions.
Because the entire ThreadLocal<T> struct will live in thread-local
storage, the ThreadLocal<T> struct can simply have a normal "T*"
inside of it to point to the value.

In order to make ThreadLocal<T> POD, I had to change it so it did
not derive from ThreadLocalBase. I also had to change the reset()
method to be non-virtual.

Because ThreadLocal<T> no longer has an 'm_key' member that it
calls pthread_getspecific() and pthread_setspecific() on, we need
some way to free the data held by the ThreadLocal when a thread
dies. I solved this issue by creating a ThreadLocalManager class
that holds a pointer to the top of a linked-list of ThreadLocals.
Each ThreadLocal<T> will add it self to the linked-list the first
time it it accessed. When a thread dies, the pthread library will
call ThreadLocalManager::OnThreadExit, which will traverse the
linked-list and free each ThreadLocal.

Because ThreadLocal<T>::reset() is no longer virtual, we need
some way for ThreadLocalManager::OnThreadExit to know which
version of the reset() method to call on each ThreadLocal. I solved
this by adding a function pointer 'm_on_thread_exit_fn' to
ThreadLocal<T> that points to the correct version of the static
method ThreadLocal<T>::OnThreadExit, which in turn calls the
correct version of the reset() method.

Finally, putting the "__thread" keyword in front of ThreadLocal
declarations and definitions broke how RequestLocal<T> used to work.
Thus, I took the technique I applied to ThreadLocal and applied the
same technique to RequestLocal.

If everyone is happy with this scheme, then I think we should extend
this technique to ThreadLocalCreate<T>, ThreadLocalSingleton<T>, and
ThreadLocalProxy<T> as well.

Test Plan:
1) make fast_tests
2) make slow_tests
3) run flib tests
4) Run www in production, verify it doesn't cause problems

DiffCamp Revision: 118910
Reviewed By: hzhao
Commenters: bmaurer, qixin
CC: hphp-diffs@lists, bmaurer, davidrecordon, hzhao, achao, qixin,
andrewparoski
Revert Plan:
OK