Rework stack overflow detection.
Summary:
This changes the stack overflow detection code to use actual stack space,
rather than a preset limit on the number of function calls. The available
stack size is obtained from pthreads, and does not need to be heuristically
determined.

This reduces the amount of code generated in function injections. Since the
old mechanism had to decrement its counter at the end of the function,
register pressure is also decreased.

Test Plan:
Built and ran www/ to ensure no false positives. Also built a test program
that infinitely recurses to check that it is properly caught.

DiffCamp Revision: 124253
Reviewed By: iproctor
CC: hphp-diffs@lists, iproctor
Revert Plan:
OK