Fix foreach to update the array's internal pointer
Summary:
This revision fixes two bugs. The first issue is that a normal foreach loop
is supposed to set the array's internal pointer before the each iteration
to point to the element after the current element. To solve this problem, I
changed ArrayIter so that it has two modes: DefaultMode and PhpForeachMode.
When we want ArrayIter to update the array's internal pointer, we can
switch it into PhpForeachMode by calling feRef() or fePtr().

Originally, I just wanted to change ArrayIter to always update the array's
internal pointer. However, I grepped through the codebase and saw that we
use ArrayIter internally in a lot of places where we probably don't want to
modify the array's internal pointer. Thus, by having two modes, we can look
through the codebase and update various uses of ArrayIter on a case by case
basis. Similarly, we use iter_begin and iter_advance a lot internally to
iterate over arrays. We should also looks at these parts of the codebase
and see if any of these loops should be changed to update the array's
internal pointer.

The second issue was that calling "key($arr)" or "current($arr)" in PHP
triggers the array to be copied if it's refcount is greater than 1, while
in HPHP it does not. This might not seem like a problem at first, but
because foreach loops can modify the array's internal pointer without
triggering the array to be copied, it actually matters when the array copy
is triggered.  For example:

  <?php
  $arr = array(0,1,2,3);
  reset($arr);
  var_dump(key($arr));    // Outputs "int(0)"
  foreach ($arr as $v) {
  }
  var_dump(key($arr));    // Outputs "NULL"
  $arr = array(0,1,2,3);
  reset($arr);
  var_dump(key($arr));    // Outputs "int(0)"
  foreach ($arr as $v) {
    if ($v == 1) { key($arr); }
  }
  var_dump(key($arr));    // Outputs "int(2)"
  $arr = array(0,1,2,3);
  reset($arr);
  var_dump(key($arr));    // Outputs "int(0)"
  foreach ($arr as $v) {
    if ($v == 1) { $arr[3] = 999; }
  }
  var_dump(key($arr));    // Outputs "int(2)"

When the first foreach loop begins, it increments the ref count of $arr.
During each iteration, it updates $arr's internal pointer without
triggering copy on write. When the first foreach loop ends, it decrements
the ref count of $arr. At this point $arr's internal pointer points past
the last element. Thus, the call to "key()" after the first foreach loop
returns NULL.

When the second foreach loop begins, it increments the ref count of $arr.
At the beginning of the first and second iteration, it updates $arr's
internal pointer without triggering copy on write. Then, during the second
iteration, the body of the loop calls "key()", triggering a copy of the
array to be made. Now the foreach loop holds its own array which is
distinct from $arr. So at the beginning of the third and fourth iteration,
the foreach loop updates the internal pointer of its own array, but it
doesn't modify $arr's internal pointer. Thus, the call to "key()" after the
second foreach loop returns 2.

The third foreach loop produces the same results as the second foreach
loop. The only difference is that it triggers copy on write by modifying
$arr.

Finally, I've added some testcases to verify that we match PHP's weird
behavior.

Test Plan:
make fast_tests
make slow_tests
Browse www

DiffCamp Revision: 124683
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao, andrewparoski
Revert Plan:
OK