[Fix] no caching for StaticStrings in hphpi
Summary:

Some AST objects in hphpi (StringName and ScalarExpression) have a field of type StaticString. This is bad, because when the AST object is destructed, the StaticString is also gone, potentially leaving a dangling pointer somewhere else, as code everywhere else assumes StaticStrings never get destructed.

Unfortunately, this seems not that easy to fix without a larger change. For now, this diff stops checking the cache for those StaticStrings in hphpi, so that the interpreter does not crash, and other people can continue working with the interpreter.

Task ID: #

Blame Rev:

Reviewers: hzhao

CC:

Test Plan:

made hphpi and browsed www

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

DiffCamp Revision: 125567