fixed parse to only ignore the first line of file, if containing #
Summary:
i was trying to handle "#!/usr/bin/php", but the fix was too aggressive. It's
now only doing that for the 1st line of file. We're good to handle # in the
middle of the file now.

Test Plan:
fast_tests, slow_tests

DiffCamp Revision: 127008
Reviewed By: mwilliams
Commenters: cwalsh
CC: hphp-diffs@lists, cwalsh, mwilliams
Tasks:
#239473: hphp parse error on single line if statement containing #

Revert Plan:
OK