[Fix] Unsetting private properties takes them out of iteration and serialization
Summary:
In the interpreter, unsetting a property would just make it null. It would
still
appear in serialization and iteration. This fixes that and cleans up the
approach for
the compiler too. propExists now uses the initialized bit on the property which
is also
how o_get choses properties for serialization.

Test Plan:
TestCodeRun

DiffCamp Revision: 127614
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK