raise error when calling BuiltinClass::InstanceMethod()
Summary:
PHP allows UseClass::InstanceMethod() calls, as long as inside the method,
there is no access to $this. We used o_id == 0 to detect that. But for extension
classes, there is no "implicit static method". If it's declared as instance one,
it cannot be called statically no matter what. Added the checking to
INSTANCE_METHOD_INJECTION_BUILTIN, which was incorrectly injected as
INSTANCE_METHOD_INJECTION without "_BUILTIN" before. So that's fixed now.

Test Plan:
This script gave the same error in HPHPi now:

  <?php DateTimeZone::getName();

DiffCamp Revision: 129751
Reviewed By: qixin
CC: hphp-diffs@lists, qixin, mpal
Tasks:
call

Revert Plan:
OK