[Fix] Fix RecursiveDirectoryIterator infinite recursion
Summary:
When a RecursiveDirectoryIterator was given to a RecursiveIteteratorIteration
and was initialized
with a path that didn't exist, it caused an infinite recursion. What php does
is throw an exception
when the RecursiveDirectoryIterator is constructed when the path doesn't exist.
This change
adds that behavior.

Test Plan:
New unit test

DiffCamp Revision: 130477
Reviewed By: myang
CC: hphp-diffs@lists, myang
Revert Plan:
OK