Include unsets in dead store elimination
Summary:
For inlining to be really effective there is a lot of cleanup that needs to be
done.

replacing $a=$x;expression-using-$a;unset($a) with expression-using-$x can
reduce a lot of clutter, and constructor/destructor calls.

This code does a lot of cleanup of Expression::Context to ensure contexts are
correct, and consistent, and also fixes some type inference issues

Test Plan:
fast_tests, slow_tests, flib/unit tests

DiffCamp Revision: 131798
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK