[Fix][Leak] Make the XboxTask sweepable
Summary:
XboxTask, which is allocated by SmartAllocator  holds a refCount
to XboxTransport, which is not allocated by SmartAllocator.
Because the destructor of XboxTask is not called, the XboxTransport
is not freed ever, causing a leak when using Xbox.

Task ID: #211924

Blame Rev:

Reviewers: hzhao

CC:

Test Plan:
tested with make fast_tests and on production with API tier.
The memory usage with this patch under the production traffic seems much
more reasonable.

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

DiffCamp Revision: 131438