Fix object type inference
Summary:
Type inference for objects was broken quite badly. Fortunately, we dont often
get the chance to use the resulting type, and so we mostly got away with it.

The main issue is that Type::Coerce(Object, SpecificObject) returns
SpecificObject, making this kind of code fail in very mysterious ways:

<?php
class X {
  public $x = 1;
  }
class Y {
  public $x = array();
}
function test() {
  $x = (object)new Y;
  var_dump($x->x);
  $x = new X;
  var_dump($x->x);
}
test();

The reason for the backwards coerce appears to be that any time we see $x->foo
we coerce $x to type Object, and we dont want that to "degrade" the type if its
already been inferred as a specific object.

Im removing the bogus coerce in ObjectProperty and ObjectMethod, fixing
Type::Coerce, and fixing NewObjectExpression to always return Object, or a
specific object type.

Test Plan:
fast_tests, slow_tests, flib & unit tests

DiffCamp Revision: 132013
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK