Update system/gen
Summary:
I made some changes to the compiler which affected the generated code, but
forgot to update the generated files.

Im temporarily setting --opts=none too, because I noticed that one of the
"optimizations" seemed to cause a couple of pessimizations for this code.

Once Ive fixed the issues, I'll re-enable opts for the system files.

Test Plan:
fast_tests, inspection of the diffs

DiffCamp Revision: 131809
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK