More cleanup of redundant assignments
Summary:
A couple of new cases handled
 - if a simple variable is read, and its value hasnt been changed since it was
written, and the value written was another simple variable whose value hasnt
changed, replace this variable with the other: $a = $b; echo $a ===> $a = $b;
echo $b
 - if a variable is unset, or overwritten, and it hasnt been used since the
previous time it was set, and an object might be involved, delete it if there is
nothing observable between the assignment and the unset: $a = foo(); unset($a)
===> foo(); but $a = foo(); echo "hello"; unset($a); is still not changed.
 - if a variable is unset, or overwritten, and it hasnt been used since the
previous time it was set, and an object might be involved, and the rhs of the
original assignment is available, we can kill the original assignment: $a = $b;
<no mods of b, no use of a>; unset($a) ===> $a = null; ...; unset($a); We can
then typically eliminate the $a = null assignment, and the unset.

Test Plan:
fast_tests, slow_tests, flib/unit tests.

DiffCamp Revision: 133629
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao, mwilliams
Revert Plan:
OK