Opening invalid file still results in a close call to -1
Summary:
Noticed this in valgrind while testing another bug.

When an invalid file is opened you get a warning and the fd is set to -1

In the destructor we never check that m_fd is >= 0.

Test Plan:
The falling run through hphpi with valgrind no longer prints a warning

<?php

$f = fopen('./invalid-file', 'r');

DiffCamp Revision: 133435
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
Ok