Clean up ';' usage
Summary:
This addresses incidental icc warnings by making
';' usage consistent (never doubled).  The approach
here is that any list of statements that always
ends with a ';' has the ';' outside of the macro,
e.g. FORWARD_DECLARE_CLASS, so that it is clear a
list is being terminated.  Macros that don't necessarily
end with a ';', e.g. IMPLEMENT_SMART_ALLOCATION_NOCALLBACKS,
do not have a ';' after the macro invocation.  In this
case it's because it defines methods which might end
with '}'

Task ID: #

Blame Rev:

Reviewers: andrew.paroski

CC:

Test Plan: it compiles, fast_tests

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

DiffCamp Revision: 136436