Collapse several start-of-function checks into a single "surprised" flag
Summary:
There are a number of conditions that can cause an exception to be thrown at
the beginning
of a PHP function: timeout, out-of-memory, signal, and it's possible that more
may be added
in the future. Instead of checking them all independently, this patch makes a
single
"surprised" flag that is set whenever any of those conditions happen.

This provides a small code size and speed improvement.

Test Plan:
slow_tests, flib tests

I don't think those tests regularly invoke the behavior that this
patch affects, though. What's the best way to exercise these codepaths?

DiffCamp Revision: 138901
Reviewed By: andrewparoski
Commenters: qixin, mwilliams
CC: hphp-diffs@lists, iproctor, qixin, andrewparoski, jpotter, mwilliams
Revert Plan:
OK