Remove extra temps for NewObjectExpression
Summary:
Now that we create a shared-pointer for the initial new expression, we should
use that for the expression's m_cppTemp when it needs one, rather than creating
a new one, and assigning the original to it (see the diffs in reflection.cpp for
what Im talking about).

Test Plan:
fast_tests, slow_tests.

DiffCamp Revision: 139560
Reviewed By: qixin
CC: hphp-diffs@lists, qixin
Revert Plan:
OK