[Perf] Pass in statically known context
Summary:
For a lot of cases, we statically know the context for accessing an object
property. More specifically, if we are in a user-defined function or class, we
know the accessing context is either empty or the current class.

This diff generates the code that passes in the accessing context to avoid one
extra thread-local access.

Test Plan:
microbenchmarks showed about 1/4 reduction in property access time

fast_tests
slow_tests
running in production

DiffCamp Revision: 142047
Reviewed By: mwilliams
CC: hphp-diffs@lists, mwilliams
Revert Plan:
OK