compressed logic still wrong
Summary:
The compressed logic is still wrong. Sending content in compressed mode is
an optimization and needs to have both the browser and the server data
agreed upon. In all other cases, we should send the uncompressed data.

Test Plan:
In the browser, with/without encoding, access a static resource that is
compressed/uncompressed.

DiffCamp Revision: 142055
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK