[Fix] Get rid of ClassStatement::nameString()
Summary:
Storing a ref-counted String in ClassStatement is not safe, even with the fix
in , because the ref count might drift
during rollback, which will eventually corrupt the memory when the server is
running long enough.

Now it removes the nameString() method and the m_nameString field, but rather
creates the String (with AttachLiteral), before passing it to FrameInjection or
assigning it to the cached class name in an object. This is safe because (1)
frames have very short life spans, and the string stored in stacktraces will get
copied, and (2) the object's life is shorter than the class'.

Test Plan:
fast_tests

running www in hphpi (debug build)

DiffCamp Revision: 142379
Reviewed By: iproctor
CC: hphp-diffs@lists, iproctor
Revert Plan:
OK