Clean up array key access with null string
Summary:
Null as an array key should be treated as "". This case has been taken care of
if the key is passed in as a Variant. This diff does the same things for keys
passed in as Strings:

Also changes String::toKey() to return the empty string for null string,
similar to Variant::toKey().

Test Plan:
fast_tests
slow_tests

DiffCamp Revision: 144040
Reviewed By: hzhao
CC: hzhao
Revert Plan:
OK