Rebuild system files after merge
Summary:
A last minute merge of Fast Method Call needed a rebuild of the system files,
which didn't happen.  This fixes that.

Test Plan:
fast_tests.  This was already tested with the Fast Method Call
check-in.  All this does is add hphp_set_error_page to system files,
and that is checked directly by running

DiffCamp Revision: 144138
Reviewed By: iproctor
CC: hphp-diffs@lists, iproctor, bill
Tasks:

Revert Plan:
OK