[Fix] line number for undefined array indexes
Summary:
For rvalAt() not in ExistContext, the line number needs to be set, so that when
a notice is raised for an undefined array index, the error handler can get the
correct position in the code.

Add a new testcase as well.

Test Plan:
test/test TestCodeRun TestErrorHandler

DiffCamp Revision: 144785
Reviewed By: hzhao
Commenters: mwilliams
CC: hphp-diffs@lists, hzhao, qixin, mwilliams
Tasks:
flib/core/utils/arrays.php on line 0)

Revert Plan:
OK