[Fix] Nested array comparison
Summary:
Two arrays could be incomparable, i.e., ==, <, >, <=, and >= all evaluates to
false. When such two arrays are nested, the nesting arrays can no longer assume
their elements are always comparable (that is, one of ==, <, and > is true).

This diff adds a new method ArrayData::equal() for == and ===, and evaluates
"<=" and ">=" for arrays as "< or ==" and "> or ==". Performance-wise, this
makes <= and >= for arrays more expensive, but == cheaper, which could be a good
trade-off, as people probably don't compare arrays with <= or >= that much, (I
guess).

Also add new testcases for nested arrays.

Test Plan:
Hanson's testcase.

fast_tests
slow_tests

DiffCamp Revision: 146500
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao, hansonw
Tasks:
#312847: Nested array equality inconsistency

Revert Plan:
OK