Retire the LiteralString type from Variant's list of types
Summary:
Though the LiteralString type initially allows us to avoid having to
allocate a StringData, it has several major downfalls. First, we often end
up calling strlen multiple times because we don't have a place to cache the
length of the string. Second, we sometimes construct a temporary StringData
on the stack (see the cast functions inside builtin_functions.cpp and
type_variant.cpp). Finally, in several places we end up having to allocate a
StringData (see calls to toString() in type_variant.cpp).

Now that allocation is relatively cheap (due to ThreadLocal and other
improvements), we get equivalent performance and less code complexity if we
simply take the initial hit of allocating a StringData. In the future, if
there are performance-critical parts of the runtime where we want to avoid
the cost of allocating a StringData, we can take advantage
of our superior StaticString infrastructure.

The shootout benchmark got a little faster (11.9s -> 11.8s), and there was
no noticeable effect on performance or peak memory consumption when I ran
www in production.

Test Plan:
1) fast_tests
2) slow_tests
3) Run www in production

DiffCamp Revision: 145999
Reviewed By: myang
Commenters: qixin, mwilliams, hzhao
CC: hphp-diffs@lists, hzhao, qixin, andrewparoski, myang, mwilliams
Revert Plan:
OK