[Fix] Fix compilation error due to unused temporary.
Summary:
This code would result in a compilation error:

<?php
function bar($g) { return $g; }
class X {
  static function foo() {
    echo $this->baz(bar(1), bar(""));
  }
}

because temporaries were output for the results of the two bar calls, but the
temps were not used, because the call to baz was replaced with a call to
throw_fatal.

Instead, use GET_THIS_ARROW(), which will throw the fatal, and allow the rest
of the expression to be output. A later diff will optimize this (and other)
throw_fatal cases.

Test Plan:
slow_tests (including above test case). Build www.

DiffCamp Revision: 153051
Reviewed By: hzhao
CC: hzhao, hphp-diffs@lists
Revert Plan:
OK