[Perf] ArrayData::add()
Summary:
This diff adds new methods to ArrayData: add() and addLval(), which are similar
to set() and lval(), but with an additional assumption that the key is not
present in the array, i.e., a new key is added. Therefore, there is no need to
do a find(). They are useful for array copy, escalation, unserialization, etc.

Test Plan:
fast_tests
slow_tests
production tests

DiffCamp Revision: 157073
Reviewed By: myang
CC: qixin, myang, hphp-diffs@lists
Revert Plan:
OK