Leases in HPHP
Summary:
The only bit that was needed was the idl.php changes

Expose the memcached lease operations to the PHP code stack through the phpmcc
module.

We add four new memcache operations:
$mcc->lease_get($k);
returns
array("value" => <value in cache>,
         "lease_token" => <lease token to pass on set>,
         "is_stale" => <indication if value is stale>,
         "is_error" => <indication if result was an error>
         "result" => <hit, miss, hotmiss>)

$mcc->lease_get_multi(array($k1, $k2));
returns an array whose kesy are the input keys and the value is an array of the
form for lease_get

$mcc->lease_set($k, array("value" => <new val>, "lease_token" => lease token
from get))
returns true if it was stored or stored as stale. false otherwise

$mcc->multi_lease_set()
takes an array whose keys are the keys to store and the values are the array of
the format above

Test Plan:
I ported the memcached leases tests into php and made sure that the new
aggressive lease test and all other libmcc tests pass

DiffCamp Revision: 148756
Reviewed By: rmcelroy
Commenters: hzhao, ttung
CC: hphp-diffs@lists, rajeshn, rmcelroy, hzhao, ttung, mpal, memcache-dev@lists
Revert Plan:
OK