[Fix] Fix various auto-vivification and side effect issues with object properties
Summary:
The rules for unset were wrong (in both hphp and hphpi, but in different ways).
In a chain such as unset($a->b->c->d), if a property exists, and is an object,
but the next property in the chain doesnt exist, it gets created (assuming its
not the last). Note that this doesnt continue up the chain because the newly
created property is null (and so not an object).
The rules for when a parameter could be CVarRef were wrong, missing various
"deep" contexts
In hphpi, unset/weakLval didnt evaluate all the side effects if the expression
turned out not to exist.

<?php
class Y {}
class X {
  public $a;
  function __construct() {
    $this->a = array('x' => new Y);
  }
  function bar() {
    var_dump('bar');
    $this->qq = new Y;
    $this->qq->x = $this->qq->y = 1;
    return $this->qq;
  }
}
function foo() { var_dump('foo'); return 'foo'; }
function test($x, $a, $s) {
  $t = &$s->t;
  unset($x->bar()->x);
  unset($x->q->r->s->${foo()});
  unset($x->y->a->b->c);
  unset($x->a['x']->y->a->b->c);
  unset($a['a']['y'][foo()]);
  unset($a['b']->y->z);
  unset($a->c->d);
  var_dump($x, $a, $s);
}
test(new X, array(), false);

Test Plan:
fast_tests, slow_tests

DiffCamp Revision: 156539
Reviewed By: iproctor
CC: iproctor, hphp-diffs@lists
Revert Plan:
OK