[Fix] More order of evaluation issues
Summary:
For chains consisting of ArrayElementExpression and ObjectPropertyExpression
nodes, the order of evaluation is: evaluate all side effects left to right, then
evaluate the chain. In addition, for assignment and assignment-op expressions,
the chain on the lhs is not evaluated until the rhs has been fully evaluated
(but the side effects of the lhs are still evaluated first).

This just addresses hphp order of eval. The fixes for hphpi will be more
complicated.

Test Plan:
Added a test case (disabled under hphpi). fast_tests, slow_tests, flib_tests

DiffCamp Revision: 158724
Reviewed By: andrewparoski
CC: achao, andrewparoski, mwilliams, hphp-diffs@lists
Revert Plan:
OK