fixed variable display problem
Summary:
The real problem is I swallowed unserialization error, so it appeared as no
variable was defined, but in fact everything was fine. This only happens to a
Serializable class that has stronger unserialization requirement, throwing when
class cannot be found. Fixed it so for debugger's unserialization, we allow an
unknown Serializable class.

Earlier gdb showed Array res is empty, and that was confirmed it's a gdb/gcc
bug :-)

Also be more forgiving on IP block map config error that stopped my debug
build.

Test Plan:
able to show local variables even when it has unknown Serializable
SimpleFBObjectProfile class

DiffCamp Revision: 164893
Reviewed By: mwilliams
CC: mwilliams, hphp-diffs@lists
Revert Plan:
OK