[Fix] chdir and error logging for light processes
Summary:
A couple of things: (1) When chdir fails in popen, log a warning instead of
failing the popen, this is because cwd in the compiled version might not exist;
(2) Log more errors in the case that popen fails; (3) Switch the order that
light processes and log file initialize, so that the light processes get the
correct log file; (4) Check whether ShellExecContext::exec() returns null to get
rid of a possible segfault.

Test Plan:
Changed cwd in the execution context to something that didn't exist, and
shell_exec didn't fail.

DiffCamp Revision: 166277
Reviewed By: andrewparoski

Revert Plan:
OK