Optimize dependency graph.
Summary:
Optimizaitions by lcastelli@gmail.com

"Profiling the hphp compiler shows that a large amount of time is spent
in DependencyGraph::add, because of the search done on the list of
Dependency objects. This patch converts the list to a map, to make the
search run in O(log(N)) instead of O(N).

According to what exactly is being compiled this seems to give from a
moderate (10%) to a huge (>100%) speedup in the parsing and inferring
types phases."

Test Plan:
fast_tests, slow_tests. build www (observed a 25% improvement in
type-inference, and 5% in parsing). build www-dev and run unit tests

DiffCamp Revision: 169278
Reviewed By: hzhao
CC: hzhao, platform-diffs@lists, hphp-diffs@lists
Revert Plan:
OK