Optimize HphpArray and fix various bugs
Summary:
The primary objective of this revision was to improve the performance of
HphpArray. This was accomplished by doing the following:

1) Currently HphpArray stores the original (possibly unaligned) data
pointer in m_data and it repeatedly computes the aligned pointer on demand.
I changed HphpArray to instead store the aligned pointer in m_data and to
store the amount that was added to make it aligned in a new field
m_dataPad. When we need to realloc or free, we can subtract m_dataPad from
uintptr_t(m_data) to obtain the original pointer value.

2) Right now HphpArray keeps track of the size of its hashtable and the
number of slots for elements using the field m_lgTableSize, which stores
the log base 2 of the hashtable size. Whenever we want to know the size of
the hashtable, the hashmask (size of hashtable minus 1), or the number of
slots for elements, we perform these computations on demand based off of
m_lgTableSize. I changed HphpArray to instead store the hashmask in
m_tableMask, and I did away with the m_lgTableSize field. Now we compute
the size of the hashtable and the number of slots for elements based of the
hashmask.

3) Almost all places where we used the Variant machinery have been changed
to use the TypedValue helpers instead.

4) The quadratic probe logic was rewritten to avoid using multiplication.

5) A new helper findForNewInsert() was added to HphpArray for situations
where we can guarantee that a given key is not already present in the
array. This is handy for when we need to rebuild the hash. We are also able
to take advantage of it for append() as well.

6) I noticed that gcc generates worse machine code when dealing with 4-byte
integers, so I changed HphpArray to use 8-byte integers in a few key places
to work around this.

7) Finally, I did some refactoring to make iter_advance faster in the
common case where there are very few tombstones.

In addition to these performance improvements, this revision also updates
all of the array implementations to handle the case where the next
available integer key overflows the same way that PHP 5.3.1 handles it.
This bugfix makes it possible for us to safely optimize append() because
we can always safely assume that the next available integer key is not
already present in the array.

Test Plan:
1) fast_tests
2) slow_tests
3) flib
4) Single machine testing

DiffCamp Revision: 179919
Reviewed By: je
CC: je, andrewparoski, hphp-diffs@lists
Revert Plan:
OK