Optimize printRaw
The overhead of appending some output to a stream is substantially
reduced in case a single large string is appended all at once compared
to appending its characters one at a time.

This patch thus optimizes the no-indent case of CodeGenerator::print to
count the newlines separately and then appending the whole string
together. Despite the need to do two passes over the string, this
approach is faster overall.

The optimization is especially useful in case PregenCPP has been
selected, since in that case printRaw gets heavily used.