Optimize CodeGenerator::print
This patch improves the previous optimization of the print function so
that it also applies in cases where indent is true. It also slightly
changes how indentation is handled which fixes some rare cases where
indentation is currently wrong in the output.

Overall I've measured about a 30% time decrease in the CPP generation
phase for a large codebase with this optimization.