[Perf] Reduce code bloat for dummy object creation
Summary:
We are duplicating a lot of code for generating a dummy object in case of
calling an instance method without $this. The majority of the compiled code of
those i_ and ifa_ functions is the dummy object generation, and that case should
be rare.

This diff adds createDummy() for each class, and replaces the duplicated code
with a call to createDummy(), reducing the binary size.

Test Plan:
fast_tests
slow_tests
production tests

DiffCamp Revision: 174217
Reviewed By: mwilliams
CC: qixin, mwilliams, hphp-diffs@lists
Revert Plan:
OK