Move forward class declaration from forward class headers to class headers
The macros for forward declarations are also useful to declare the
p_##cls typedefs which are not declared anywhere else. Thus it makes
sense to have a forward class declaration of a class in its header
file, and just leave the forward header for the forward declarations
needed to satisfy the dependencies.

This patch also removes FORWARD_DECLARE_REDECLARED_CLASS since that
seems unused.