Add full class file inclusion for header files
When a class constant expression calls the lazy_initializer of the
corresponding class, its header file must have been included before (a
simple forward declaration is not sufficient, since a method is
called). This is not correctly done when the constant expression is a
default parameter, so it gets expanded in a header file. This patch
covers that case.