Add class constant declarations for header files
This adds class constant declarations to forward headers.

In addition the replacement of a class constant with its value is
removed from outputCPPImpl. This could create dependency problems since
the value could be an expression dependent on other declarations, which
would have not been added in the analysis/optimization phases. Note
that in case the value is static the replacement is already done during
the pre-optimization, so this additional occasion for replacement is
not needed in that case.