Avoid mallocing for the common case for foreach by reference
Summary:
HipHop's array implementations use the PointerList structure defined in
"util/pointer_list.h" to keep track of all the strong iterators that are
currently iterating over a given array. PointerList was designed to take up
minimal space and avoid mallocing memory when there are no elements.

This revision further optimizes PointerList so that it avoids mallocing
memory until a second strong iterator is added to the PointerList (which
happens very infrequently). This is accomplished by storing the first
strong iterator in the PointerList's m_data field and marking the low bit.
All relevant PointerList operations now check the low bit of m_data first.
If the m_data is non-NULL and the low bit is not set, we know m_data points
a malloced array of pointers to strong iterators. If the low bit is set, we
know that the (m_data & ~1) points to the first and only strong iterator in
the PointerList.

This revision also changes HipHop's array implementations to raise a
warning instead of throwing a fatal error when unsetting the next element
during foreach by reference.

Task ID: #

Blame Rev:

Reviewers: hzhao, qixin, je

CC:

Test Plan:
make fast_tests
make slow_tests

Revert Plan:

Tags: hphp

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

DiffCamp Revision: 175226