[Fix] SetCallingObject needs to get root
Summary:
Found a problem with late static binding and MethodCallPackage, while doing the
FrameInjection optimization. See the added test case.

Also, most of the flags in CallInfo didn't seem to get set. This diff starts to
set StaticMethod to reduce the number of times of calling SetCallingObject() and
getRoot(), but CallMagicMethod is still not used.

Test Plan:
TestLateStaticBinding

DiffCamp Revision: 175044
Reviewed By: iproctor
CC: iproctor, hphp-diffs@lists
Revert Plan:
OK