[Un-perf] Rewrite preOptimize step to avoid redundant processing
Summary:
Previously, we iterated full passes of preOptimize until nothing was done. This
resulted in 6 passes for our code base.

This change creates a dependency graph, and processes BlockScopes in bottom up
order, and only re-processes as necessary. The result is that each BlockScope is
optimized 1.3 times on average.

The overhead of the dependency management ends up being quite large (and can
probably be improved), but we still get a 2.5x speedup.

The change is fairly large, so Im checking in just the preOptimize re-org. I
will follow up with postOptimize, and then type-inference.

I also discovered (and fixed) many minor bugs, with the result that a number of
class constants are no longer inlined (the classes are volatile, but the values
were being inlined before we realized that) - hence the Un-perf tag. On the
bright side, I see a way to eliminate/optimize a lot of existing
checkClassExists testing, which will probably offset any losses.

Test Plan:
fast_tests, slow_tests, build www-dev and manually inspect diffs in code (the
differences all correspond to fixed bugs, as far as I can see). flib/unit tests.

DiffCamp Revision: 177897
Reviewed By: myang
CC: myang, hphp-diffs@lists
Tasks:
#402203: speed up hphp compilation

Revert Plan:
OK