pdo::fetch() and fetchAll() didn't respect a default fetch type
Summary:
While the manual shows PDO::FETCH_BOTH as the default parameter its really
PDO::FETCH_USE_DEFAULT which is 0 and its checked and set to the class default.

PHP doesn't expose FETCH_USE_DEFAULT so had to hardcode this as 0 in the idl so
we don't get a parse error. Is this oK?

Test Plan:
<?php

$pdo = new PDO('sqlite::memory:', '', '');
$pdo->exec('CREATE TABLE foo (bar STRING)');
$pdo->exec("INSERT INTO foo VALUES ('ABC')");
$result = $pdo->query("SELECT * FROM foo");
$result->setFetchMode(PDO::FETCH_NUM);

foreach ($result as $row) {
  var_dump($row);
}

Revert P
Ok

DiffCamp Revision: 176290
Reviewed By: hzhao
CC: hzhao, hphp-diffs@lists
Revert Plan:
OK