better way to generate global_variables.h
Summary:
The code change wasn't that simple, as I had to change "gv_" to "gvm_" and
"ci_" to "cim_" on GlobalVariables, because we code generated "gv_" and "ci_"
for local or class constants. So, now they are replaced with GV() and GCI().

Test Plan:
fast, slow, www

DiffCamp Revision: 178964
Reviewed By: mwilliams
Commenters: bmaurer
CC: bmaurer, mwilliams, hphp-diffs@lists
Tasks:

Revert Plan:
OK