[Fix] Don't wait for ever on proc_close()
Summary:
We use waitpid() to close lightly forked child processes, but waitpid() doesn't
have a timeout, causing some threads to block for ever.

This diff uses RequestTimeoutSeconds as the timeout for
LightProcess::waitpid(), and sends a SIGKILL to the child process once the
timeout is triggered.

Test Plan:
fast_tests with LightProcess turned on

Used light processes to proc_open() a binary that never ended. It was correctly
killed after the timeout.

DiffCamp Revision: 180512
Reviewed By: hzhao
CC: hzhao, hphp-diffs@lists
Tasks:
#433090: Kill spawned processes after the request timeout has happened

Revert Plan:
OK