[Perf] Strong type hints
Summary:
When a function is declared with type hints, but it can be dynamically invoked,
we made the parameter Variant, rather than the specific type.

This change creates wrapper functions so that we can have inner functions with
strongly typed parameters, and wrapper functions with Variant params.

It also fixes a few bugs where we didnt properly type check the params. In
particular, params with default values were not type checked, and hints of
unknown, or redeclared classes were not type checked.

Unfortunately, php allows you to use an error handler to bypass the type hint
checks, so this code is controlled by an option: HardTypeHints. When the option
is set, the function will not be invoked regardless of what the error handler
returns (the function body will not be executed, and the function will return
null).

Test Plan:
fast_tests, slow_tests, flib/unit tests.

DiffCamp Revision: 181238
Reviewed By: hzhao
CC: hzhao, hphp-diffs@lists
Revert Plan:
OK