Add support for sort_flags in array_unique
This adds support for a second parameter in array_unique added starting
from PHP version 5.2.9. SORT_STRING/SORT_LOCALE_STRING use the already
existing implementation, SORT_NUMERIC uses a similar approach, while
SORT_REGULAR replicates the algorithm implemented in PHP more closely
in order to reproduce exactly the same behavior. The PHP approach tends
to be slower especially in case most of the elements in the input array
are duplicates, so it is used only when necessary.