[Fix] Call-time pass-by-ref could leak the contagious bit
Summary:
If a function parameter is determined to be CVarRef (because no modifications
are made to that param), but a callsite passes that param explicitly by
reference, nothing consumed the contagious bit, and the next assignment of that
param was binding (so really bad things could happen).

In this example, changing $b changes the passed in parameter:

<?php
function f($x) { $b = $x; $b++; }
$a = 1;
f(&$a);
var_dump($a);

Fix it by setting the lvalParam bit on the parameter's symbol, so its not
implemented as CVarRef.

Also add a new bit, callTimeRef, so that alias analysis can do a better job
with non-reference parameters that dont have the callTimeRef bit set (not yet
used).

Test Plan:
fast_tests, slow_tests

DiffCamp Revision: 183693
Reviewed By: qixin
CC: qixin, hphp-diffs@lists
Revert Plan:
OK