Improve performance of HPHP::VariableUnserializer
Summary:
Replaced an inefficient loop for copying data between
buffers with a call to memcpy.

Test Plan:
Tested the code in a sandbox, as well as a standalone
code that was extracted into a test utility.

Also did fast_tests and slow_tests and production run.

DiffCamp Revision: 184159
Reviewed By: andrewparoski
CC: hzhao, mberezecki, andrewparoski, hphp-diffs@lists
Tasks:

Revert Plan:
Revert a commit.