Fix race condition for EntityMap initialization
There is a race condition in the lazy initialization of the EntityMap:
in case the map is not initialized and threads A and B are just before
the initialization code, the following can happen:

             A              |              B
---------------------------------------------------------
 !EntityMapInited == true   |
 Lock EntityMapMutex        |
 !EntityMapInited == true   |
 Set EntityMapInited = true |
                            | !EntityMapInited == false
                            | Use EntityMap

In this case thread B uses the map before or while it is initialized.
Even moving the instruction that sets EntityMapInited to true after the
initialization is not safe without a memory barrier because of possible
reordering by the optimizer or the CPU.

This patch thus moves the initialization of map to be done by a static
constructor at startup time, which seems simpler and safer.