[Perf] Reduce number of temporaries
Summary:
There are a lot of places where we generate a temporary for the result of an
assignment, which already has a temporary for the rhs:

 t1 = foo;
 t2 = (x = t1);
 t3 = (y = t2);
 ... t3 ...

We can drop the extra temps in that case:

 t1 = foo;
 x = t1;
 y = t1;
 ... t1 ...

Test Plan:
fast_tests, slow_tests, flib tests.

Size of stripped www-dev dropped by 1M (not a big deal - but this was obviously
quite common).

DiffCamp Revision: 188083
Reviewed By: qixin
CC: qixin, hphp-diffs@lists
Revert Plan:
OK