Make htmlspecialchars_decode decode only special characters
Currently the implementation of htmlspecialchars_decode always decodes
all html entities. This might not always be desiderable, especially
since the function doesn't accept a specification of what encoding to
use, and thus it also always assumes UTF-8.

This patch changes the function to make it more similar to PHP and
decode only special characters. This way the fact that no encoding can
be specified is not a problem, since special characters have the same
encoding in both ASCII/latin-1 and UTF-8.