Limit the size of xhprof trace buffers.
Summary:
Add an option to limit the size of xhprof trace buffers.  As
the code stands (and by default) trace buffers will extend to cover
the all calls in a request.  In the event of a truncated trace,
resources accessed after the trace terminates will be recorded in
a special arc, allowing the effect of such events to be noted in
resulting data.

Test Plan:
Run on a sandbox with a small trace buffer configured,
observe that the trace truncates properly.

DiffCamp Revision: 188565
Reviewed By: ps
Reviewers: ps
CC: ps, hphp-diffs@lists
Revert Plan:
OK