[hotfix] disabling s_exit_callback
Summary: My bad on code review and I shouldn't allow this code to get checked in :-) s_exit_callback has to be on ExecutionContext that's thread local as a very minimal requirement to not crash. In fact, it should be request local.

I'm just commenting out the line, so we can stop crashes on dev boxes. Charlie, you can come up with a full fix.

Reviewers: qigao, qixin

Test Plan: no crash on dev boxes

DiffCamp Revision: 189308