[Fix/Perf] Fix array lvalues in object context
Summary:
For $a['x']->foo = 'bar', we would optimize the lvalAt by not copying the array
if the element existed. But if the element existed, and wasnt an object, we
could then end up modifying a shared array (even a static array).

<?php
function test() {
  $a = array(false);
  $a[0]->foo = 1; // oops we just modified a static array
}
test();
var_dump(array(false));

In this case, we modified the static array 'array(false)', and the var_dump
prints out an array containing a stdClass object.

To fix that, we need to check that the element exists and is an object, or is a
referenced variant. By adding the second check, we can now apply the same
optimization to $a['x']['y'], and so potentially avoid more array copies than
before.

Test Plan:
fast_tests, slow_tests. Manually tested the new test case with HphpArray and
ZendArray. Using SmallArray faults on startup (with or without my change).

DiffCamp Revision: 192832
Reviewed By: qixin
Reviewers: qixin, je, andrewparoski
CC: qixin, hphp-diffs@lists
Revert Plan:
OK