Add support for calls from eval to compiled functions
This patch adds the last bits of code to support calling a compiled
function from eval, when eval support is enabled.

For example, compiling the following code with "EnableEval=1" now
works:

<?php
function test() {
	echo "A", PHP_EOL;
}
eval('test();');
?>

Support for calls to redeclared functions is supported as well using
the CallInfo structure.

The patch also fixes a couple of issues related to casing and reference
counting in eval.