[Fix/Perf] atomic Countable and AtomicString
Summary:
Currently, Countable, together with StringData and Variant, has to live within
a single thread, unless it is set to static. However, we sometimes do want to
share objects between threads, while allowing them to be destructed, (unlike
StaticStrings or ScalarArrays, which are never destructed).

This diff adds another flag to Countable, and those that mimic Countable, which
is the "atomic" flag. For example, to make an atomically referenced StringData:

  StringData *s = new StringData(...); // Cannot be NEW(StringData)(...)!
  s->setAtomic();

Now, this StringData object can be safely passed to a normal SmartPtr, e.g.,
String, because incRefCount() and decRefCount() that are called from String are
ignored, once setAtomic() is called on the StringData object.

This StringData object can still be ref counted, but with incAtomicCount() and
decAtomicCount(), which are just like incRefCount() and decRefCount(), but are
thread-safe. An AtomicString wrapper is added to serve as the "SmartPtr" for
atomic StringData. It could be further generalized to be some kind of
AtomicSmartPtr, but I would rather do it when there is a real demand to do so.

AtomicString is applied to a bunch of ASTs in the interpreter, so that we can
use Strings safely, especially with fibers.

Not in this diff, but the ClassInfo const char * => String change is added
back.

Test Plan:
fast_tests
slow_tests
flib tests
production tests
/scripts/unittest/tests/platform/graph/batch.php

DiffCamp Revision: 196495
Reviewed By: hzhao
Reviewers: hzhao, qigao, myang
CC: hzhao, qixin, hphp-diffs@lists
Revert Plan:
OK