Allow for permanently locking code memory
Summary:
Faulting the entire binary into memory provides us with
predictable performance, but the initial thoughts about allowing the
binary to be paged out later if memory is needed is proving to be a bad
thing and reduces performance when the binary is paged out.  Default to
munlocking the binary.

Test Plan: production run
Reviewed By: je
Reviewers: mwilliams, qigao, je
CC: jasont, ps, mwilliams, je
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 291027