[Fix] Dont report missing functions in derivedFromRedeclared classes
Summary:
We reported an error for unknown functions in derivedFromRedeclared classes,
but we dont know for sure whether the function exists or not.

Test Plan:
build www-dev and verify that the incorrect error from the task has gone.
(fast_tests and www are fubar, so no point in trying to test them).

DiffCamp Revision: 197436
Reviewed By: hzhao
Reviewers: hzhao
CC: hzhao, hphp-diffs@lists
Tasks:
#463463: Nemo fails to track __construct (or all methods?) across parent/child
class chains (?)

Revert Plan:
OK