[Fix] Dont generate uncompilable files on parse errors
Summary:
When a file has a parse error, the filescope is still created, but with no
pseudoMain. We would then generate calls to the pseudoMain, which wouldnt
compile.

This changes things so we dont put an entry in dynamic_table_file.cpp, and we
dont generate direct calls to such files. This results in a runtime error if you
try to include the file, rather than a compile time error.

Ideally, we would generate a meaningful runtime error (rather than file not
found) - but I'll add that as a bigger change later (also handling a variety of
other error conditions).

Test Plan:
build www and www-dev with no compile errors.

DiffCamp Revision: 197449
Reviewed By: hzhao
Reviewers: hzhao
CC: hzhao, hphp-diffs@lists
Revert Plan:
OK