[php53] removed __xhp_idx and fixed reported yield bugs
Summary:
Changed "yield;" to "yield break;" to be really explicit. "yield null;" has to
be written explicitly as "yeild null;" not "yield;", which is illegal.

Fixed a problem with public property access on a system class, when it has to
be Variant.

Test Plan:
fast_tests, slow_tests, www and this script,

  <?php
  function foo() { return array(1, 2, 3);}
  var_dump(foo()[2]);

DiffCamp Revision: 198570
Reviewed By: mwilliams
Reviewers: mwilliams
Commenters: zrait
CC: mwilliams, zrait, hphp-diffs@lists
Tasks:

Revert Plan:
OK