[Fix] Fatal when a directory is require'd
Summary:
hphpi only checked for existance of a required file, not whether it was a
directory, so it didnt fatal when given a directory name.

Fixing this so we dont have to wait until a change goes to production to find
the error.

Test Plan:
fast_tests

DiffCamp Revision: 198823
Reviewed By: hzhao
Reviewers: hzhao
CC: hzhao, jiqingt, vinodv, hphp-diffs@lists
Tasks:
#466712: hphpi doesn't fatal on a require of a directory

Revert Plan:
OK