Add looping service threads
This adds a new kind of service threads that executes a document in a
loop. It can be useful over normal service threads with an infinite
loop coded in PHP in that memory is collected after each execution of
the document. This allows running a task indefinitely and periodically
recover objects that formed reference cycles by restarting the
execution of the document.

Since the document is executed cyclicly this kind of services is not
fit for initialization tasks that must be executed before the webserver
is started. Therefore they never block the startup of the webserver,
and hphp_service_thread_started is a no-op.