Get rid of static s_effectsTag
Summary:
s_effectsTag was used to minimize the cost of computing getContainedEffects
(and hence hasEffect).

The problem is that its a global, making it hard to parallelize things that use
it.

This diff moves it into the containing scope (typically, a function scope), and
also resolves some existing issues that were mostly masked by the fact that any
change anywhere would cause the cached effects to be recomputed.

This will make it easier to parallelize preOptimize and postOptimize

Test Plan:
fast_tests, slow_tests. No change to generated files for www-dev.

DiffCamp Revision: 205857
Reviewed By: myang
Reviewers: myang
CC: myang, hphp-diffs@lists
Revert Plan:
OK