Make OOM a catchable fatal
Summary:
The OOM exception is not really a hard error that we have to immediately stop
the request for. This diff makes it a catchable normal FatalException, so that
an error handler can catch it and log more information.

The way we detect the OOM is actually already preventing recursive OOM
exceptions.

Test Plan:
fast_tests

Small scripts which releases the memory and then uses more in the error
handler. The OOM is thrown and caught only once.

built www, and run it on dev machine

DiffCamp Revision: 206537
Reviewed By: ps
Reviewers: ps
CC: ps, hphp-diffs@lists
Tasks:
#481264: Make OOM a catchable exception

Revert Plan:
OK