Make JobQueue optionally waitable
Summary:
The JobQueue object has a "stop" function, which waits until all queued jobs
are done. But it prevents new jobs from being enqueued. This makes it hard to
write code where the jobs themselves can submit new jobs.

In the parser, we wrote a loop where we would create a new jobqueue, submit all
the known jobs, and wait until they were done, recording a new list of files to
parse. Then we would repeat.

This was inefficient, because we had to keep creating and destroying threads,
and we had to keep a list of pending jobs, and we had to write much more code
than necessary to manage the jobs.

This adds an optional parameter to the JobQueueWorker which declares it
waitable. You can then call waitEmpty on the JobQueue, and it will run until all
jobs have completed.

Test Plan:
fast_tests slow_tests, minimal change to generated files for www-dev (df_lambda
functions and file-names changed - but should now be stable regardless of parse
order).

DiffCamp Revision: 206925
Reviewed By: qixin
Reviewers: qixin
CC: qixin, hphp-diffs@lists
Revert Plan:
OK