Add some missing options in cmake and put them in a separate file
The options about infinite loop, infinite recursion and request timeout
detection were defined in the cmake files but not passed to the
compiler. This patch adds them in case they are set to ON.

In addition the patch adds some other options, which are left to OFF by
default (except for ENABLE_LATE_STATIC_BINDING). Finally all the
options are put together in a file Options.cmake.