[Fix] Fix fiber marshaling problem with SharedMap
Summary:
When marshaling array parameter for fiber, we get ref of
everything in the array and set them to the array in fiber thread.
However, the "get ref" itself can change the original array when the
array is a SharedMap, e.g., creating a StringData wrapping around
SharedVariant and put it into cache.

The problem here is that the marshaling happens in the fiber thread, so
all thread-local allocations belong to fiber, while the parameter array
that gets modified belong to the regular worker thread. Then later on
both threads will try to destroy the objects that get created during
marshaling. Fiber does so because it wants to sweep everything, while
worker also wants to cleanup the thread-local array.

This diff makes the fiber marshaling going through copy() and makes
copy() using the getValue() , which does not modify (or create)
localcache for SharedMap. Basically, it avoids getValueRef() from being
used for SharedMap in this cross-thread setting.

Test Plan:
make fast_tests
manually verify with the the bug triggering request

DiffCamp Revision: 210765
Reviewed By: myang
Reviewers: qixin, myang
CC: trvish, myang, hphp-diffs@lists
Tasks:
#486706: curl request to graph batch API causes hphpi to segfault

Revert Plan:
OK