[Perf/HPHPi] Reduce calls to realpath
Summary:
In RequestEvalState::includeFile we call realpath() to get a canonicalized
file path (resolving symlinks). Under the hood realpath() calls lstat for
each component and therefore involves a lot of lstat() calls. It was reported
that the cost of calling lstat() can be as high as 5% of CPU time when loading
the home page.

Before calling realpath, File::TranslatePath is called and therefore the
translated path is already in form of an absolute path, with possible symlink
component(s).

I added a runtime option CheckSymLink that is turned off by default. When
turned on, we falls back to old way of calling realpath. When turned off,
if the path is already in absolute, we consider it as already resolved and
skip the call to realpath.

This does affect __FILE__ so if a symlink is involved in the translated
absolute path, then HPHPi replaces __FILE__ with a different string from
that of PHP. After doing some test, this semantics difference does not seem
to affect our www code and flib/unit tests.

Task ID: #

Blame Rev:

Reviewers:
mwilliams, qigao
CC:
hphp-diffs@lists.facebook.com
Test Plan:
make fast_tests
flib test
unit test
run hphpi sandbox mode and browse the site

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

DiffCamp Revision: 211385