[Fix] Preserve self-reference for array copy
Summary:
This is a different way of fixing the array copy-on-write problem, when an
ArrayData has one strong reference from one of its elements, and another weak
reference from outside.

When the array is copied, we can check whether the element is a strong
reference to the array that is being copied, and preserve the reference, even if
it is the last strong reference.

The fix is applied to ZendArray, SmallArray, and HphpArray to pass the new
TestReference testcase.

Test Plan:
fast_tests
slow_tests
flib tests

DiffCamp Revision: 212222
Reviewed By: mwilliams
Reviewers: andrewparoski, mwilliams
CC: qixin, mwilliams, hphp-diffs@lists
Tasks:
#488792: hphp doesn't serialize/unserialize references properly

Revert Plan:
OK