HPHP: Update to ICU 4.6
Summary:
This is the hphp side of the externals diff I sent out to add
ICU 4.6:

In this diff, I update the hphp build to use the new version of ICU.  I
enabled a few unit tests that were commented out, added a unit test
to confirm the fix for Task #491834, and removed a test whose behavior
changed in the new version of ICU.

Drew and I found that the behavior of ICU has changed slightly for one
of the tests that compares the Unicode collation API's metadata
output against PHP's.  We decided the change was intentional, and removed
that particular test.

While I was testing, I found a bug in ICU's spoof implementation and filed
ICU bug #8341:

http://bugs.icu-project.org/trac/ticket/8341

I can work around this on the HPHP side, but we'll probably want to work
around it on the PHP side since we'll need to do normalization there
anyway.

Test Plan:
Pointed external-centos and external-centos-dev to my externals
build.  Ran fast_tests.

Differential Revision: 212418
Reviewed By: andrewparoski
Reviewers: andrewparoski, mwilliams
CC: iproctor, bagashe, beng, andrewparoski, hphp-diffs@lists
Tasks:

Revert Plan:
OK