HPHP: Link to ICU dynamically for .so builds
Summary:
This is the other side of D212630.  I re-enabled linking
dynamically to ICU for the .so builds of the HPHP runtime.

Technically, we could link against the .a libraries, except that makes
the slow_tests target take a long time.  (Also, the .a libraries are
not built with -fPIC, which I forgot about.)

Test Plan:
fast_tests, slow_tests (in progress)

Differential Revision: 212634
Reviewed By: qixin
Reviewers: qixin, andrewparoski, myang
CC: arice, hphp-diffs@lists
Revert Plan:
OK