[Fix] Fix refcounting issues in derived-from-redeclared classes
Summary:
Inside cloneSet for a derived-from-redeclared class, we would clone the parent,
and then set it, even though we had already constructed a parent.

This is inefficient, but also caused an issue when the parent had a __destruct
method; the frame-injection for the __destruct method would create an object for
the root object, which has (at this point) a refcount of zero. So when we exit
the __destruct method, we free the root object (and free the parent object a
second time), and then we return the free'd root object from clone.

Even absent clone, there is a similar issue for release on a
derived-from-redeclared object. The root object's release method is called,
which causes the parent object's release method to get called, which causes the
__destruct method to be called; and again the root object is passed to the
frameinjection, which increments, the refcount from 0 to 1; so on exit from
parent->__destruct, the frameinjection destroys the root object a second time.

Additionally, EvalObjectData implemented clone by doing a clone on the
EvalObjectData itself, followed by a clone of the parent. If the parent had a
parent, there was no way to set the grandparent's root pointer.

Test Plan:
fast_tests, slow_tests flib/unit tests. Specifically

Added a test case:
  <?php ;
  if (isset($g)) {
    class X {}
  } else {
    class X {
      public $a = 1;
      function __destruct() {  var_dump(__METHOD__); }
    }
  }
  class X1 extends X {
    public $t = 1;
  }
  function test() {
    $x = new X1;
    $x->t = 5;
    $x->a = 3;
    $y = clone $x;

    var_dump($y);
  }
  test();

DiffCamp Revision: 214620
Reviewed By: qixin
Reviewers: qixin
CC: qixin, mwilliams, hphp-diffs@lists, cbueno
Revert Plan:
OK