Fix possible crash in case jemalloc is used without stats support
Summary:
From Hyves:

This modifies the memory manager to avoid crashing in case jemalloc is
used but it was not built with stats gathering support. It also makes
it use weak references to the mallctl* functions, for uniformity with
what is done elsewhere in the code.

Test Plan:
Built hphp
check stats still works

DiffCamp Revision: 214279
Reviewed By: mwilliams
Reviewers: ps, mwilliams
CC: mwilliams, hphp-diffs@lists
Revert Plan:
Ok