[Fix] "instanceof self" and "instanceof parent"
Summary:
The instanceof operator needs to handle self and parent. In the compiled
version, self and parent are replaced with the real class names, and in the
interpreter, FrameInjection is used to fetch the names.

New testcases are also added.

Test Plan:
fast_tests
slow_tests

DiffCamp Revision: 215153
Reviewed By: mwilliams
Reviewers: mwilliams
CC: bill, qixin, mwilliams, hphp-diffs@lists
Tasks:
#499470: zend/hphp inconsistency: "instanceof self"

Revert Plan:
OK