Refine work on dynamic constant
Summary:
I reworked on whether a constant is dynamic by adding checkDefined(). If a
constant is defined but replaced at postOptimize time, it will be marked
as "replaced". A constant can become undefined due to dead code elimination,
in this case it will not have the "replaced" flat set. At anytime when a
constant becomes undefined because of optimization, it is marked as dynamic.
I also made a fix for the replacing value which can be false if user is
trying to define a system constant. Undefined constants are no longer issued
and the literal strings for them are used instead.

Test Plan:
make fast_tests
make slow_tests
flib/unit test
make www and inspect the generated code.

DiffCamp Revision: 216446
Reviewed By: mwilliams
Reviewers: mwilliams
CC: mwilliams, hphp-diffs@lists
Revert Plan:
OK