Make the Variant default constructor faster
Summary:
This revision introduces a new type called "KindOfUninit" to represent
uninitialized null, and it updates the runtime accordingly. With this
change, KindOfNull now only represents initialized null values, and we no
longer set m_data.num to 0 or 1 to distinguish between uninitialized null
and initialized null. The benefit of this change is that now the default
constructor for Variant only has to zero out the _count and m_type fields,
which is can do with a single 8-byte write to memory.

This revision also removes the KindOfByte and KindOfInt16 types to reduce
the number of cases for switch blocks in Variant logic and other parts of
the runtime.

This change reduces cpu time for www in production by approximately 0.7%,
and it reduces the www binary size by 1.9%.

Test Plan:
make fast_tests
make slow_tests
Build www, run in production, measure perf and verify logs look normal
Flib tests under hphpc and hphpi

Differential Revision: 215912
Reviewed By: qixin
Reviewers: qixin, myang
Commenters: bmaurer, myang
CC: bmaurer, qixin, andrewparoski, myang
Revert Plan:
OK