Fix end() to clear the IterationDirty flag
Summary:
HPHP has an IterationDirty flag inside ZendArray to detect when foreach
loops are mixed with current(), each(), key(), value(), prev() and next()
so that it can raise a notice.

Currently we are clearing the IterationDirty flag when reset() is called,
but we are failing to clear it when end() is called. This revision fixes
end() to correctly clear the IterationDirty flag.

Test Plan:
make fast_tests
Write a PHP program to exercise this and verify the fix works

Differential Revision: 218459
Reviewed By: qixin
Reviewers: qixin, mwilliams
CC: qixin
Revert Plan:
OK