[Fix/yield] yield in anonymous function in hphpi
Summary:
The synthesized variable $__cont__ in the generator function might get a wrong
index, because it could be synthesized out of the function, and for anonymous
functions, there might be an outer function, in which the $__cont__ gets a wrong
index. This only matters for the interpreter.

This diff adds a new method onSynthesizedVariable(), which is essentially the
same as onSimpleVariable(), but doesn't generate the index for hphpi.

New tests are added and passed.

Test Plan:
TestYield

DiffCamp Revision: 219712
Reviewed By: mwilliams
Reviewers: mwilliams
CC: mwilliams, hphp-diffs@lists
Tasks:
#505179: Enable yield inside of anonymous functions

Revert Plan:
OK