[Fix/hphpi] file inclusion with cwd == "/"
Summary:
The relative path of a file is computed from cwd, if SourceRoot is empty.
However, Util::relativePath() gives wrong result if a '/' is blindly added to
the end of the cwd, making it "//", when cwd is "/".

This diff fixes that by only adding '/', when there is no '/' at the end of the
server_root.

Test Plan:
(thanks to epriestley for setting up the testing environment).
built hphpi and modified php_wrapper to call my own hphpi binary
replaced the hphpi in /home/engshare/devtoolstest
successfully committed a diff in a directory checked out with https

fast_tests

hphpi sandbox

DiffCamp Revision: 220172
Reviewed By: mwilliams
Reviewers: myang, mwilliams
Commenters: myang
CC: epriestley, ender, myang, mwilliams, hphp-diffs@lists
Tasks:
#505643: HPHPi Commit hooks are failing on some machines

Revert Plan:
OK