Fix string to int conversion cache
Summary:
From Hyves

The code currently treats the first bit of the string hashes as a cache
of whether a conversion to a numeric was attempted and failed. The code
in general expects the bit to be set in case a conversion to a numeric
was done with allow_errors=1, but the bit is actually set even in case
allow_errors was 0.

For example the following snippet:
<?php
$a = '1A';
$firstTest = ($a == 0);
$a == '0';
$secondTest = ($a == 0);
var_dump($firstTest == $secondTest);
?>

prints false under hiphop while of course it should be true, since
$firstTest and $secondTest are the result of the same computation.

The problem is that the comparison between $a and the string '0' tries
to do a strict conversion of $a to a numeric, which fails and is
erroneously cached. The second comparison between $a and the integer 0
should instead do a loose conversion to a number which would succeed,
but it skips it completely because the cache bit has been set in the
previous operation.

Test Plan:
make fast_tests
make slow_tests

Reviewed By: myang
Reviewers: qixin, myang, mwilliams
Commenters: qixin
CC: qixin, myang
Revert Plan:
Ok

Differential Revision: 226330