Change StringBufferLimitException to derive from ExtendedException
Summary:
A case was reported where a StringBufferLimitException was being thrown by
the runtime but the resulting fatal error message did not provide a file
name, line number, or backtrace. The root cause of the issue appears to be
that StringBufferLimitException derives directly from Exception instead of
deriving from ExtendedException, so StringBufferLimitException does not
capture a stacktrace when it is constructed, and so handleFatalError() does
not have a stacktrace to work with when it produces the fatal error
message.

This revision changes StringBufferLimitException to derive from
ExtendedException so that it captures the stacktrace at the time the
exception is thrown.

Test Plan:
make fast_tests
make slow_tests
Build www and run in prod

Reviewed By: qixin
Reviewers: hzhao, qixin
Commenters: ahupp
CC: qixin, ahupp, andrewparoski
Tasks:
#514396: hphp StringBuffer limit fatal does not print useful stack

Revert Plan:
OK

Differential Revision: 223619