Add 'hphp_is_service_thread()' and improve error messages
Summary:
From Hyves:

This adds a function to check whether the current thread is a service
thread or not.

Also in case hphp_service_thread_started or hphp_service_thread_stopped
are called from a thread that is not a service thread, they currently
fail in ThreadLocalProxy with the error "ThreadLocalProxy::get() called
before set()". This patch adds a check to make the error message more
to the point.

Reviewed By: mwilliams
Reviewers: qixin, myang, mwilliams
CC: mwilliams
Revert Plan:
Ok

Differential Revision: 230961