[Fix] array_filter should pass params to the filter func by value
Summary:
We were passing the result of secondRef. Its never safe to hold a long-term
reference to that, but in particular, if the underlying array is a SharedMap,
hphpi will try to box the SharedVariant, leading to later crashes.

Test Plan:
fast_tests, the typeahead game described in the task

Reviewed By: qigao
Reviewers: qigao
Commenters: tomocchino
CC: tomocchino, qigao
Tasks:
#513285: HPHP segfault in version d301236-2011-03-10

Revert Plan:
OK

Differential Revision: 224854