Fix hphpi's semantic checks for the __construct method
Summary:
This diff updates the semantic checks performed by hphpi to deal with the
__construct method correctly. When a class extends an abstract class, the
signatures of the __construct methods do not have to be compatible; however
when a class implements an interface, the signatures of the __construct
methods must be compatible.

This diff also updates the semantic checks performed by hphpi so that two
signatures are considered incompatible if one returns by value and the
other returns by reference.

Task ID: #

Blame Rev:

Reviewers: qixin, myang, qigao

CC:

Test Plan:
make fast_tests

Revert Plan:

Tags: hphp

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 225132