[Fix] Fix crash in smart allocator
Summary:
My previous change caused a crash in smart allocator because when loading
the runtime option config file, it calls preg_match which has String
arguments. We may be able to use std::string to replace String but the
change is bigger. I moved the function InitAllocatorThreadLocal back
to its original place which is before RuntionOption::Load(). Then as soon
as RuntimeOption::EnableMemoryManager is loaded, if it is false, we reset
MemoryManager accordingly.

Test Plan:
make fast_tests
make slow_tests
make www and browse the site

Reviewed By: qixin
Reviewers: qixin, qigao
CC: qixin, hphp-diffs@lists
Revert Plan:
OK

Differential Revision: 226001