Hash pointers.
Summary:
Using pointers' bitwise values as their hash is really
unacceptable. The low-order bits are almost always clear, leading
to heavy collision behavior on power-of-two-sized hash tables.

The only runtime use of note right now appears to be serialization code
for arrays, which probably is not too critical. Still, it made Jason
and me nervous to see this in the code base.

Test Plan:
1605/1605 on the make tests. Reviewers: what else?

Reviewed By: je
Reviewers: jsobel, hzhao, qixin, je, mwilliams
Commenters: mwilliams, qixin
CC: kma, qixin, je, mwilliams, hphp-diffs@lists
Revert Plan:
OK

Differential Revision: 227094