Add some live range analysis
Summary:
This adds a live range analysis pass, and a variable coalescer on top of it.

The interference analysis part of this is currently very conservative: live
ranges that coexist within a block are considered to interfere. This could be
improved with an extra pass over each block. It may be worth doing that with
inlining turned on, since inlining tends to produce a lot of very short lived
(sub block) variables.

Task ID: #

Reviewed By: qigao
Reviewers: myang, qigao
CC: arice, qigao
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 235582