[Fix/Stable] Unable to find X_map in building separable extensions
Summary:
After we switched to no -fPIC, we start seeing this problem.
Looks like there is a typo in the makefile for separable extensions,
which cause the object list to be empty. That will lead do dummy
schema.so and finally cause we cannot find X_map in hphp when loading
it. This diff fixes the type so that we can get the actual .pic.o file
linked into schema.so

Test Plan:
build separable extensions and check the symbols in schema.so

Reviewed By: qixin
Reviewers: qixin, myang
CC: qixin, qigao
Tasks:
#519398: "Unable to find ..._map in ..." while building separable extensions

Revert Plan:
OK

Differential Revision: 228170