Fix PID file race condition
There is a slight race condition regarding the creation/removal of the
PID file when a new hiphop instance is started using the takeover
functionality. In particular in certain cases it's possible that the
old instance removes the PID file after the new instance has already
created the new PID file. In this case the old instance will actually
remove the PID file of the new instance, resulting in a situation where
Hiphop is running but there's no PID file on the file system.

This fixes the problem by removing the PID file synchronously before
the new instance creates the new PID file.