Eliminate __set_state from hphp's hierarchy.
Summary:
PHP's docs call __set_state a "magic method", but the only
thing magical about it is that the strings generated by var_export call
it when eval'ed. We can tolerate this by simply generating code for
static __set_state normally.

NB: this diff is against the release branch, and the bug it fixes is
holding up the push.

Test Plan:
make tests failed 1 test: test_ext_datetime, which
fails in an unmodifed tree as well.

I need help with the test plan here; in a perfect world, this would go out
into production ahead of the Tuesday push. If this is impossibly
aggressive, please tell me so!

Reviewed By: qixin
Reviewers: iproctor, hzhao, qixin, mwilliams
CC: qixin, hphp-diffs@lists
Tasks:
#526164: LATEST is broken with HPHP/thrift error

Revert Plan:
OK

Differential Revision: 229451