Fix crash in ServerStats::collect
Summary:
We were seeing somewhat frequent crashes on the API tier due to an issue
with memory corruption.

The root cause was that sometimes a thread can exit and free the
ServerStats object associated with it, but we still have a pointer to the
ServerStats object in the ServerStats::s_loggers vector. (I'm still in the
process of investigating exactly how a thread can exit when the server is
in a steady state).

This revision fixes the issue by updating ServerStat's destructor to remove
the ServerStat object from the s_loggers vector before the object is freed.

Reviewed By: kma
Reviewers: kma, mwilliams
CC: kma
Revert Plan:
OK

Differential Revision: 230926