[Perf] Optimize VarNR
Summary:
VarNR has a no-op destructor. But gcc isnt smart enough to optimize it away. So
we end up generating a test and call twice for each one (once for the normal
flow, and once for stack unwinding).

Change VarNR to a class with no destructor, and a CVarRef operator instead.

With this done, using VarNR is always at least as good as Variant, even for
non-refcounted types; and often avoids the destructor cost. So extended
VarNR parameter passing to include scalars, and primitive types.

Reviewed By: qigao
Reviewers: qixin, qigao, myang
CC: qigao, mwilliams
Revert Plan:
OK

Differential Revision: 232477