Fix crash in PDO with error thrown in parameter casting
Summary:
The stmt parameter of the PDOBoundParam class wasn't cleared.

When an exception was called in the underlying function that normally
set that parameter it would segfault during destruction.

In this case it was string casting that failed and resulted in a fatal error.

Task ID: #527731

Blame Rev:

Reviewers: mwilliams, kma

CC:

Test Plan:

Revert Plan:
Ok

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 232197