[Fix] Fix sigma build
Summary:
sigma build seems to be ignoring the dependency  "${HEADER} : ${CLS}"

I've never seen ${} used as makefile syntax before, but it seems to work for
me. But changing it to $() to see if thats the issue.

Also changing the definitions of CLS and SRC to avoid executing those shell
commands multiple times.

Reviewed By: kma
Reviewers: kma
CC: stein, astrehl, kma
Revert Plan:
OK

Differential Revision: 233127