[Fix] Pagelet server needs to support is_uploaded_file/move_uploaded_file
Summary:
These two PHP functions did not work in the pagelet server because the
rfc1867UploadedFiles (in upload.cpp) belongs to the original request
thread and in the pagelet thread that set is empty. I made changes to
pass along the rfc1867UploadedFiles (a set of uploaded file names) to
the PageletTransport and made two virtual functions so that
transport->isUploadedFile and transport->moveUploadedFile will work
when the transport is a PageletTransport.
Task ID: #

Blame Rev:

Reviewers:
qigao
CC:
hphp-diffs@lists.facebook.com

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 234236