[Fix] Fix File::readRecord for stream_get_line
Summary:
We did not check eof when doing File::readRecord, which could
lead to returning junk data and seg fault. This diff fixes this bug as
well as another bug in the function where m_writepos could be
overwritten. It also changes the behavior in eof case to return ""
instead of false, which matches PHP better.

Reviewed By: myang
Reviewers: myang
CC: myang
Tasks:

Revert Plan:
OK

Differential Revision: 235052