[Fix] Fix hphpi(hphpd) crash when eval a closure
Summary:
When using a closure in the debugger, it caused crash
hphpd> =print_r(array_map(function($a) { return $a;}, array(1, 2, 3, 4)));
Core dumped: Segmentation fault

The issue is that we need a EvalFrameInjection instead of a plain
FrameInjection in order to pass the closure object. In this case the debugger
created a plain FrameInjection for psuedo main which won't work, causing
an assertion to fail when compiled under debug mode and crash in no-debug
mode. I made the change to create a EvalFrameInjection and associated it
with the environment.
I also fixed a read-beyond-buffer when the name of the debugger pseudo main
is just "_".

Task ID: #

Blame Rev:

Reviewers:
mwilliams, qigao, hzhao
CC:

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 235980