OpenSSL might not comes with SSL2 protocol
Under debian testing (libssl1.0.0d-1), the SSL2 protocol might be
disabled in libssl by removing the corresponding methods:

From openssl/ssl.h:

  #ifndef OPENSSL_NO_SSL2
  const SSL_METHOD *SSLv2_method(void);       /* SSLv2 */
  const SSL_METHOD *SSLv2_server_method(void);    /* SSLv2 */
  const SSL_METHOD *SSLv2_client_method(void);    /* SSLv2 */
  #endif
----

This patch makes hiphop mimic behaviour of openssl by not using this
methods and failing to setup the socket. A warning is raised to inform
the developer/user.