[Fix] Fix slow test failure
Summary:
The TestHint in the slow test could pass when running alone but would fail
as part of the entire suite. I found that the GetTypeHintTypes will populate
the TypeHintTypes map just once when first time referenced and it is empty.
This won't work in case of doing slow test since when
Option::EnableHipHopSyntax is set, the TypeHintTypes wouldn't be empty so
the rest of the types are not set in the map, causing the compiler to believe
that 'int' is an object type name.

I added a new function ResetTypeHintTypes and call it before running each
individual test.

Task ID: #

Blame Rev:

Reviewers:
qigao, mwilliams
CC:

Test Plan:
make fast_tests
make slow_tests
Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 237320