IPv6 support
Summary:
When calling getaddrinfo to bind & listen on INADDR_ANY,
0.0.0.0 should not be used if IPv6 support is asked for.  Pass NULL in
to evhttp_bind_socket_backlog_fd for the address to bind to which will
allow getaddrinfo to return an AF_INET6 address if the host supports it.

Changes to libevent are needed to support AF_INET6 sockets which is
coming in a separate diff.

Task ID: #

Reviewed By: qigao
Reviewers: qigao
CC: macvicar, ps, hphp-diffs@lists, qigao
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 237856