[Fix] Respect the --force option
Summary:
With --force enabled, hphp is supposed to keep going when it finds errors, and
generate code that fatals at runtime. For files specified on the command line,
if the file fails to parse (due to syntax errors, or because the file doesnt
exist), we dont respect the flag, and just fatal.

Task ID: # 543886

Reviewed By: je
Reviewers: je, qigao
Commenters: qigao
CC: boris, hphp-diffs@lists, ps, je, qigao
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 239349