[Fix] Parse dynamic variables in encaps lists correctly
Summary:
When a dynamic variable appears in an encaps list, we have to treat a
ConstantExpression specially: ${x} outside an encaps list means the variable
with the name of the constant x, but ${x} inside an encaps list means the
variable with the name "x". However, in an encaps list, when the name wasnt a
constant expression, we bailed and created a variable with a null name,
resulting in a runtime assert, rather than dropping through to the default
behavior.

<?php
$x = 'stuff';
$stuff = 1234;
echo "${$x}";

Test Plan:
fast_tests, slow_tests (including new test).

Reviewed By: qigao
Reviewers: qigao
CC: michalburger1, ps, qigao
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 243201