[Perf] Improve Array::set/setRef/add
Summary:
Array::set etc were implemented as inline functions calling a template
implementation function. Since the implementation function was defined in a
header (variant.h), it too could have been inlined - which would have been bad
from a code size point of view. Looking at the generated code, it seems that gcc
did not inline the template (a good thing).

But this still meant we had instantiations of the template for bool, int, int64
and double, when we only needed one, and that gcc had to instantiate the
template in every file that used Array::set* (although the linker should unify
the defs). Finally, since the template is declared to take a (const T &) the
compiler had to generate a stack based temporary for the bool, int, int64 and
double cases (at the call site) which wasnt needed.

I moved the template definitions into type_array.cpp, and marked them
ALWAYS_INLINE. This makes the code at callsites smaller (no stack based
temporary), the fast path in the function is shorter, and there are fewer
distinct instantiations of the template (so each one is "hotter" than before,
and more likely to be in the cache).

At the same time, I removed the last references to the "contagious" bit (some
asserts that it wasnt set), and removed the weakBind class, since weak binding
will happen automatically.

Test Plan:
fast_tests slow_tests

Reviewed By: qigao
Reviewers: qigao, myang
Commenters: myang
CC: ps, mwilliams, qigao, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 252358