Fix for task 560277
Summary:
Change the generated invoke_file to first check to see if there is a compiled
version of the file before attempting to eval it. This fixes the issue reported
and makes EnableEval=2 more useful. There are still some advanced scenarios
around redeclaring a class within an eval that will not work.

For example:

if (false) {
  class Foo {}
}

eval ( class Foo {} );

Will result in an error indicating that class Foo is being redeclared while it
actually isn't. Mark and I discussed this and we agreed that this wasn't worth
fixing at this time.

Test Plan:
-Validated that the repro in the task now works fine with EnableEval=2.
-Inspected the generated invoke_file code to make sure it was emitted properly.
-Tested the changes with Option::GenHashTableInvokeFile set to false.
-Ran all the unit tests (fast + slow tests).
-Verified that using a class that's compiled by hphp from inside an eval works
properly.
-Finally, did some sanity testing on the site running on my sandbox with the
updated hphpi.
-I considered adding a unit test for this; however since it's fairly unlikely
to get broken in the future and it's not critical to FB, I decided it wasn't
worth the effort and additional test execution time. Mark, if you think I should
add a unit test for this, please let me know.

Reviewed By: mwilliams
Reviewers: mwilliams
Commenters: macvicar
CC: cliffchang, ps, mwilliams, macvicar, dmortenson
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 252167