[Perf] Generate Scalar Variant
Summary:
When a literal value is used in a context (e.g., function argument) where CVarRef is expected, we currently construct a VarNR and uses it as a CVarRef. We can avoid the construction of VarNR if we pre-build all such VarNRs for literal values, much in the same way we compute scalar arrays or static string. I added a runtime option UseScalarVariant that when turned on will just do that. I also added things like true_varNR, false_varNR, etc. for a few constants. For integers/doubles, the generated VarNRs are made read-only by dumping out the contents of those VarNRs and put them in a const int64 array. I haven't done the same for literal string/scalar array VarNRs but it should be possible.

Task ID: # 531768

Blame Rev:

Reviewers:
qigao, mwilliams
CC:
hphp-diffs@lists
Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 252590