[Perf] Cleanup array functions
Summary:
We arranged a while back that Variant, String, and Object would have compatible
layout, so that we could cast a Variant that was known to hold an array directly
to Array&, rather than constructing a new array. We've not made much use of
that, so this diff replaces most uses of toArrNR with a suitable accessor.

Also, there were a lot of inlined functions in ext_array.h. Originally, these
were simple wrappers around Array or ArrayUtil methods, but they have been
modified to validate their args, and throw exceptions as necessary (in order to
match php behavior). This no longer seems reasonable, and is simply causing
bloated code.

Test Plan:
fast_tests slow_tests

Reviewed By: myang
Reviewers: qigao, myang
CC: ps, mwilliams, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 254094