[Perf] Implement one level deep CSE for ArrayElementAccess nodes
Summary:
In order to try and reduce the time spent in ArrayData::get(),
this commit implements common sub-expression elimination for
a simple, yet common case of array access. Now, code such as:

  if ($x[0]) do_something_with($x[0]);

is effectively transformed into:

  $t = $x[0];
  if ($t) do_something_with($t);

except that a C++ reference is used, so we do not have to incur
the cost of reference counting and also so we do not have to
insert explicit unsets. This optimization only applies for
array accesses of one level deep, since this is essentially
the only case we can be sure that there are no self-references.

This optimization is guarded with a new option,
ArrayAccessIdempotent, which defaults to on. By turning this flag
on, the user indicates an explicit contract with the compiler that
all ArrayAccess objects behave sensibly, that is the
offset{Get,Set} methods do not rely on side-effects for
correctness.

Note that this optimization does alter the number of warnings
generated in certain cases. That is, for code such as:

  foo($y[1]);
  bar($y[1]);

if, $y[1] is not a valid offset, then only one warning is emitted.

Reviewed By: mwilliams
Reviewers: mwilliams
CC: ps, mwilliams, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 255287