[Perf] Better code gen for generators
Summary:
The parser transformation that creates the generator marks the hints the
continuation parameter as Continuation. But that doesnt take effect, because the
body of the function contains an LVariableTable, and so "anything" could happen
to the continuation parameter.
I marked the parameter "hidden", which says to keep it out of the variable
table. This allows the compiler to deduce that its not altered, and hence that
its type is Continuation throughout the function. So now we get direct calls to
the continuation methods, rather than indirect calls.

I then made some minor tweaks to the Closure/Continuation methods to fix return
types, so that eg getLabel() returns an int, making the initial switch more
efficient.

Test Plan:
fast_tests slow_tests

Reviewed By: myang
Reviewers: qigao, myang
CC: ps, mwilliams, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 259841