[Perf] Reduce the number of destruct() and release() functions generated by user defined classes
Summary:
Instead of having each class define destruct() and
release() separately and nearly identically, move most of the
logic into ObjectData. The only logic which each user class
needs to retain is defining an operator delete overload which
releases the object back to the list allocator. Additionally,
with this change, release() on an object no longer needs to be
virtual.

Reviewed By: mwilliams
Reviewers: mwilliams
CC: mpal, cbueno, ps, mwilliams, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 262729