[Perf] Improve code generated for runtime type-casting
Summary:
This commit attempts to improve the code generated for type casting in
two ways.

First, it cleans up the ambiguous use of expected type in
Expression.cpp- if the type is set, then the type of the expression must
be converted (if necessary) to that type (which is same as the old
meaning). However, if the type is set, it now means that no type
conversion is necessary, where before the meaning was either (1) actual
type and expected type are the same, or (2) no type conversion is
necessary. Doing so allows us to distinguish when we can safely omit a
cast.

Second, it adds a full set of "fast cast" methods to variant, and
adds the necessary code generation to take advantage of the
fast casting method wherever possible.

Test Plan:
slow_tests, inspect www code

Reviewed By: myang
Reviewers: mwilliams, myang, qigao
Commenters: mwilliams
CC: ps, mwilliams, stephentu, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 268015