[HPHP Tainting] Static string detection
Summary:
- created a new taint bit, TAINT_BIT_MUTATED, for non-static string detection
- altered base.php to accept the names of constants (e.g., "TAINT_BIT_HTML")
instead of the values (e.g., 0x1) and defaulted set_mask to TAINT_BIT_MUTATED
for all functions
- made a better HPHP warning message for taints

Note that only the string.idl.php functions (and the fb_taint family of
functions) have been regenerated in the codebase to handle the new taint bit
properly.  This avoids bloating the diff; the generated code will be pushed
in a diff after this is reviewed.

Also note that the chosen mechanism marks all strings as not-mutated by
default and taints them whenever they pass through any library function.
This is because the layer in HPHP that seems to map to PHP types (various
source files in runtime/base) doesn't appear to distinguish between PHP string
literals and other PHP strings; moreover, it seemed that messing around in
the parser was a Bad Idea (TM).  If either of these things are false, please
bring it up in review.

NOTE: Moved unit test changes to a separate diff, undid changes to generated
code except for ext_fb (also all in a separate diff).

Test Plan:
- fast_tests, slow_tests
- unit tests in tests/tainting

Reviewed By: myang
Reviewers: srenfro, qigao, mwilliams, myang
Commenters: pad, amenghra, qigao, srenfro
CC: collingreene, ghotz, mwilliams, pad, mwang, amenghra, qigao, srenfro, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 268875