[Perf] Closures with use vars are now implemented as subclasses of a base Closure class, instead of using an Array to pass the environment around.
Summary:
For each closure which declares use vars, a lightweight anonymous closure class
is generated which declares the use variables as instance members on
the closure class. This allows us to more efficiently pass the
variables along, instead of storing them into a PHP array and pulling
them at upon entry of the closure.

The only exception to this is a closure which is also a generator, in
which case we fall back to what we used to do. Once we clean up
generators, then we should also be able to clean this up.

There is also additional cleanup of closures to become more consistent with
zend.
First of all, explicitly instantiating a closure will now throw a
fatal. Second of all, we now get the semantics for duplicate use
variables right. For example, consider:

  $f = function () use ($x, &$x) { $x++; };

versus:

  $f = function () use (&$x, $x) { $x++; }

In PHP 5.3, it is the last declaration of the use variable
which takes effect. So in the first case, $x is by reference,
and in the second case $x is by value. Third of all,
subclassing a closure will throw a fatal when you try to
instantiate an instance of that subclass. The reason we don't
throw a fatal on seeing the declaration is because hphpc does
not have a means to properly enforce final classes.

Reviewed By: myang
Reviewers: myang, qigao, mwilliams
CC: ps, mwilliams, stephentu, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 269343