Add debug build info to --version flag for hphpc
Summary:
Sometimes I forget whether or not the particular build of hphpc
I have is a debug build or not, and previously there was no easy
way to check. I modified --version to include this information.

Test Plan:
run with --version and see if it makes sense

Reviewed By: qigao
Reviewers: myang, qigao, mwilliams
CC: hphp-diffs@lists, ps, mwilliams, qigao
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 274954