[Fix] Allow __invoke() to be declared as abstract
Summary:
This was an externally reported bug. If __invoke() was
declared as abstract, then this resulted in un-compilable code since
the __invokeCallInfoHelper() + helpers were being generated regardless.
This diff fixes this issue by only generating the helpers if __invoke()
is not abstract.

Test Plan: slow_tests
Reviewed By: mwilliams
Reviewers: mwilliams, myang, qigao
CC: ps, mwilliams
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 277186