RuntimeOption::ThreadLoopDocumentsSleepSeconds
When implementing daemons or cronjobs as a ThreadLoopDocument
the approach of making them responsible for sleeping between calls is a bit fragile.
It easily becomes error prone in case of many documents
and might result in one thread consuming all the resources.

A RuntimeOption::ThreadLoopDocumentsSleepSeconds was added
to specify default sleep time for all loop documents.