[Pref] Type propagation was not optimally propagating specific object types
Summary:
Type::Inferred() would bail out immediately with the first operand if
both of the operands were KindOfObject. This was sub-optimal, however,
if the first operand was a standard object, and the second operand had a
specific object type. This diff fixes this sub-optimality.

Test Plan: slow_tests
Reviewed By: mwilliams
Reviewers: mwilliams, myang, qigao
CC: ps, mwilliams
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 280573