[Fix] Disallow serialization of closure or continuation object
Summary:
We decided not support serialization of closure or continuation object
for now. This diff explicitly raise fatal for that. For hphpd to work
properly, add one more serialization format which is basically normal
serialization except it uses dummy string in place of closure or
continuation object.

Test Plan:
make fast_tests
make slow_tests
tried a test file with closure, check that serialization of an
object with a field set to closure gives a fatal but hphpd can
print that object.

Reviewed By: mwilliams
Reviewers: myang, mwilliams, stephentu
Commenters: stephentu
CC: ps, mwilliams, qigao, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 281886