Change json loose to options, this is part of the first rollout
Summary:
Looks for a boolean first, and if it is then assumes loose. We can remove
this once the www changes are out.

We're not exposing the rest of the options yet, we'll work on this soon.

Task ID: #614206

Blame Rev:

Reviewers: mwilliams

CC:

Test Plan:
make fast_tests
flib tests with www change

Revert Plan:
Ok

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 283031