[Fix] Previous fix to Fiber introduced new issues
Summary:
FiberJobs were refcounted, but unconditionally free'd when the job finished.
But the FiberAsyncHandle could still have a reference to it at this point, so it
could end up trying to decref (and delete) something which had already been
freed.

I tried to fix that by making the job decRef itself when it finished; this
solved the double free issue, but was still wrong because the FiberAsyncHandle
should not free the job (since the job's smart-allocated objects were actually
allocated in the job's thread, not the FiberAsyncHandle's thread), unless it
didnt run asynchronously.

So went back to having the job delete itself, but enable the FiberAsyncHandle
to detect whether or not its appropriate to do the decRef.

Test Plan: TestCodeRunEval-Fiber TestCodeRun-Fiber, and some server mode unit
tests, with and without fibers enabled.
Reviewed By: qigao
Reviewers: qigao, myang, swapnil
Commenters: myang
CC: ps, mwilliams, qigao, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 285034