[Fix] Fix the problem hphpd stuck in code mode after done
Summary:
When we finish the code input mode, we run processEval() which changes
the inputState after running Eval. That's a problem because if we hit a
breakpoint during Eval, the inputState is not properly switched to
TakingCommand. This diff switches the order.

Test Plan:
make fast_tests
manually try a small case like the one described in the task

Reviewed By: myang
Reviewers: myang, mwilliams
CC: ps, mwilliams, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 285535
Task ID: 639232