Make defining constants to non-scalar values fail
Summary:
I recently introduced warnings for doing this, while still allowing the
constant
to be defined to the non-scalar value. This changes it so that the define is a
no-op and returns false.

Test Plan:
fast_tests, slow_tests.

Search www for blatant instances of this ("sgrep_php -e 'define(X,
array(...))'") and find that there aren't any. Any instance will have been
warning in hphpi for a couple of weeks, so it shouldn't ever be happening.

Reviewed By: mwilliams
Reviewers: mwilliams, myang
CC: hphp-diffs@lists, ps, mwilliams, oyamauchi
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 288139
Task ID: 606317