[Fix] Delay the clearing of AccessorEffect for an ArrayElementExpression until post-opt
Summary:
This is to deal with a problem where a variable gets inferred
to be either an array or string, so it has its AccessorEffect cleared,
and then later on it turns out the inference was too optimistic, so it
should have its AccessorEffect flag set again.

This diff fixes this issue by delaying clearing of the AccessorEffect
flag until post-opt.

Test Plan: slow_tests, make sure www compiles
Reviewed By: mwilliams
Reviewers: mwilliams, myang, qigao
CC: ps, mwilliams, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 290569