[Fix] Properly scope calls to i_ functions
Summary: We can sometimes call an indirect function dispatcher directly. The
prefixes are the same for both methods and global functions, so if we're calling
a global function from inside a method, we might need to qualify the global
function's name. Since its rare, and its hard to figure out when we need to,
just qualify it always.
Test Plan: fast_tests slow_tests
Reviewed By: stephentu
Reviewers: qigao, stephentu, myang
CC: edward, ps, mwilliams, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 293744