[Fix/HPHPi] Fix invalid memory reference
Summary:
FrameInjection::GetBacktrace returns an array of stack backtrace. It had
assumed that the function name is a C++ literal string or a string that
lives long enough that the reference to it is still valid when the stack
backtrace is later displayed. This assumption is wrong for HPHPi where the
EvalFrameInjection takes a composed function name that is C++ stack allocated
and can go out of scope when exception is thrown. I fixed the problem by
making a copy of the string for EvalFrameInjection in GetBacktrace.

Test Plan:
make fast_tests

Reviewed By: qigao
Reviewers: qigao
CC: hphp-diffs@lists, ps, mwilliams, myang, qigao
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 294503
Task ID: 656756