Fix synchronization issues with 'surprised' flag
Summary:
The HPHP runtime has several flags in RequestInjectionData to track if a
thread has exceeded memory usage, timed out, or received a signal. As an
optimization, RequestInjectionData also has a 'surprised' flag that is set
to true when one of the other flags is set to true. Because some of the
flags can be set by another thread or from a signal handler, the flags are
protected by a recursive mutex to guard against race conditions and to make
sure the 'surprised' flag is always consistent with the other flags.

However the mutex does not properly deal with synchronization in the face
of signal handling. When a thread has acquired the mutex and is in the
middle of reading/writing the flags, it is possible for the thread to get
interrupted by a signal, enter the signal handler, acquire the mutex again,
and read/write the flags.

This diff addresses the problem by consolidating all the flags into a single
machine word and using atomic operations to modify the flags instead of
using a mutex.

Test Plan:
make fast_tests
Run www in prod

Reviewed By: mwilliams
Reviewers: je, mwilliams
CC: ps, mwilliams
Differential Revision: 295844
Task ID: 648142