[Fix] do not abort sending response when having an invalid header
Summary:
The code used to throw an exception on invalid header, but the exception
is never handled. Since we don't actually need an exception there, this
diff changes to log the error and proceed so that the response is still
sent just without the invalid header.

Test Plan:
make fast_tests
test server with the newly added test
manually tried server mode and saw it doesn't abort.

Reviewed By: mwilliams
Reviewers: mwilliams, myang
CC: macvicar, ps, mwilliams
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 301251
Task ID: 667158