[HPHP Tainting] Augmenting warning system
Summary:
Having HPHP spit out warnings on taint may not be the most desireable
behavior in all situations. This diff adds a runtime configuration
option that can be used to suppress warnings. If warnings are
suppressed, we tally them in a request local array, which can be
retrieved using an exposed function. This lets us do things like detect
when a request has echoed HTML-tainted strings and then choose to
re-run the request with a trace enabled to get more information.

We could also store the tainted outputs but for now we don't.

Test Plan: looked at logs while toggling trace and warning
Reviewed By: myang
Reviewers: srenfro, mwilliams, amenghra, myang
CC: ps, mwilliams, mwang, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 301523
Task ID: 656564