[HPHPi] Added FB_MySQLLexer class as a built-in to speed up HPHPi unit tests.
Summary:
Lexer.php was taking 10-30% of the time with PHP unit tests. Dropped
this file inside HPHPi to make unit tests faster.

I guess, if we do it this way the file goes to the open source version
also. I looked into it, but I couldn't figure out an easy way to put
this under facebook/extensions.

Let me know if the class should be named differently or if this is not
a good way of doing this.

Test Plan:
Tested that a bunch of unit tests under flib/ were working and running
a lot faster.

Reviewed By: mwilliams
Reviewers: mwilliams, myang
Commenters: myang
CC: kbeck, test-eng@lists, achao, ps, mwilliams, tuomaspelkonen, myang
Differential Revision: 303172
Task ID: 665197