Don't run certain tests dependant on specific versions of libjpeg/png
Summary:
Some tests use the local PHP (which is linked with system libs
for jpeg and png) to generate output to compare to HPHP's output.
Two of these tests generate JPEGs and PNGs, which will be different
depending on the version of libjpeg and libpng used causing tests
for HPHP built against the gcc-4.4.5-glibc-2.11.2 platform to
fail.

Therefore, this diff causing these tests to only run if a "native"
externals repo is used, which in this case is just the default
"hphp" repo.

Test Plan:
ran the tests for gcc-4.4.5-glibc-2.11.2 and
"hphp" builds.

Reviewed By: mwilliams
Reviewers: gpatangay, mwilliams, ldbrandy
CC: ps, mwilliams
Differential Revision: 306435