Fix slow tests for gcc-4.4.5-glibc-2.11.2 platform
Summary:
The test runner programmer was still being compiled with the
centos toolchain even when the rest of HPHP was built against
the gcc-4.4.5-glibc-2.11.2 platform.  This diff changes the
makefile that builds the runner to include 'rules.mk' so that
the runner program will be built with the correct toolchain
dependending on the externals platform HPHP is built for.

Test Plan:
ran slow tests for both gcc-4.4.5-glibc-2.11.2 and
native builds

Reviewed By: mwilliams
Reviewers: mwilliams, ldbrandy, gpatangay
CC: ps, mwilliams, agallagher
Differential Revision: 306750