[Perf/HPHPi] Speed up BinaryOpExpression and ArrayElementExpression
Summary:
I did some preprocessing on BinaryOpExpression and ArrayElementExpression
to record at compile time so we can do faster fetch for common cases:
VariableExpression and ScalarValueExpression. We can avoid a virtual call
to eval, and also a Variant copy construction because the context is a
CVarRef context and the copy is not needed.
I also fixed a race condition that existed because of Variant stored in
ConstantExpression. It is replaced by ScalarValueExpression by the optimizer
and the Variant in ScalarValueExpression becomes static. Changed the rand()
call to use the pointer as random number instead to avoid some perf cost.

Test Plan:
make fast_tests
make slow_tests
flib/unit test

bench.php reduced from 39.x to 37.x.

Reviewers: mwilliams, qigao

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams

Differential Revision: 317815