[HPHP Tainting] Pass -DTAINTED to targets built w/ tainted hphp
Summary:
Failing to define TAINTED for build targets of a tainted hphp binary
will result in segfaults due to conflicting definitions. To ensure this
never happens, when we compile the hphp binary, if the build is tainted,
we place a sentinel file in HPHP_LIB; then for any build under the
tainted binary, we check for the existence of this file and pass
-DTAINTED if we find it.

In order to switch from tainted to untainted or vice versa, a clean is
needed so that we rebuild hphp with the code guarded by TAINTED flags.
This removes the sentinel file. However, since code generation is taint-
blind, we can still compile code generated under a tainted build with an
untainted build and vice versa by pointing HPHP_HOME and HPHP_LIB to the
desired repo.

Test Plan:
Ran 'make TAINTED=1 -C src -j' and made www without passing TAINTED;
ran hphpi and www as server with no crash and saw taint logs. Then did
the same for untainted builds, seeing no taint logs.

Reviewed By: mwilliams
Reviewers: mwilliams, gpatangay
CC: srenfro, hphp-diffs@lists, ps, mwilliams, mwang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 309659
Task ID: 656564