Remove use of dlopen in test runner
Summary:
Previously, to run tests, the 'run' binary would dlopen the
libtest.so shared library for the current test and run its
'main' function.  This caused problems with glibc, as it has
an issue where using pthread_join in static initializers
when run via dlopen causes a deadlock:
  - http://sourceware.org/bugzilla/show_bug.cgi?id=12307
  - https://bugzilla.redhat.com/show_bug.cgi?id=661676
The problem occurs when initializing newer versions of the
mysql library (newer versions than what is in the HPHP externals
repo, which is why the problem hasn't been seen yet), which is
used from the fbcode third-party repository.

This diff replaces the use of dlopen in the test runner to
use the dynamic linker instead.  This required building against
a dummy libtest.so, then using LD_LIBRARY_PATH to point the
test runner to the correct libtest.so when running the actual
tests.

Test Plan:
Ran slow tests for builds of HPHP
1) Against its own externals tree
2) Against the fbcode gcc-4.4.5-glibc-2.11.2 externals tree
3) With TAINTED=1

Reviewers: mwilliams, ldbrandy, gpatangay

Reviewed By: mwilliams

CC: mwang, ps, mwilliams, ldbrandy

Differential Revision: 311686