[Fix/Taint] Don't leak taint trace into APC
Summary:
Some of the new trace sites introduced in the IDL audit were causing
trace to leak into the APC, which caused segfaults when we attempted to
interpret the data in later requests. When placing data into the APC, we
copy it into a shared context, so to fix the issue we just temporarily
disable HTML tracing so that we don't pass our trace to the copies.

Extensive testing seems to indicate that this is the only such edge case
caused by the IDL audit.

Test Plan: Ran unittest suite on www with no segfaults.

Reviewers: mwilliams, srenfro, amenghra

Reviewed By: srenfro

CC: ps, mwilliams, srenfro, mwang

Differential Revision: 311549