[Perf/HPHPi] speed up AssignmentOpExpression::eval
Summary:
More than half of the LHS in an AssignmentOpExpression is a VariableExpression.
We already made a special case for that. I made a change to take advantage of
the newly available inlinable getRef so getting the variable does not go
through a virtual call of eval(env).

Test Plan:
make fast_tests
bench.php execution time reduced by about 0.5 ~ 1s.

Reviewers: qigao, mwilliams

Reviewed By: qigao

CC: hphp-diffs@lists, ps, mwilliams, qigao

Differential Revision: 311617