[Fix] parallel query doesn't retry connections
Summary:
Parallel query does a retry when the connection is successfully opened but
the query fails. It does so by reopening the connection and retry the query.
Reopening the connection involves closing the previous one if any and
re-establish the connection from scratch. However, if the first connection
fails, it makes no attempt to try reopening the connection and bails out
immediately. I made a change so that if the first connection fails and the
retryQueryOnFail is true, subsequent attempts are made to open the connection
again until success or a maximum number is reached. This may help if the first
connection failure is due to network glitch. I also made the number of
query retries controled by a similar runtime option.

Test Plan:
make fast_tests
make slow_tests

Reviewers: qigao, mwilliams

Reviewed By: qigao

CC: rthiessen, hfisk, hphp-diffs@lists, ps, mwilliams, qigao

Differential Revision: 313107

Task ID: 637591