[Fix] Deal with parse errors more cleanly
Summary:
When parsing fails, we were left with half-finished classes and methods, which
often failed to meet the expected invariants, resulting in crashes or
assertions.

Add a cleanup step to remove all the classes and functions from a FileScope
when parsing fails. A later diff can create a pseudoMain with a suitable runtime
error - for now we will just get file-not-found if you try to include a file
with a syntax error (thats the same as current behavior, except that half the
time we get compilation errors, or hphp crashes).

Test Plan: fast_tests, slow_tests. Introduced lots of syntax errors into www,
and built it. No crashes from hphp, no uncompilable files, and no linker errors.

Reviewers: qigao, myang

Reviewed By: qigao

CC: ps, mwilliams, qigao

Differential Revision: 314639

Task ID: 686329