[Perf] ObjectData::o_properties should be an Array
Summary:
For some reason it was an Array*, meaning that every use was indirected (and
checked), in addition to the smart pointer indirect and check.

This should be entirely good, since sizeof(Array) == sizeof(Array*) - so we're
simply eliminating a level of indirection and checking, with no downside.

Test Plan: fast_tests slow_tests

Reviewers: myang, qigao

Reviewed By: myang

CC: ps, mwilliams, myang

Differential Revision: 317713