[Fix] Fix the posix_getrlimit
Summary:
ArrayData::set(litstr, ...) will not copy the string, the string there
is allocated on stack so by the time it is used for printing, it
contains garbage data. This diff added a wrapper for that to make sure
the string is valid.

Test Plan:
make fast_tests
try posix_getrlimit manually

Reviewers: mwilliams, myang

Reviewed By: mwilliams

CC: ps, mwilliams

Differential Revision: 316134

Task ID: 694960