[HPHPi] Allow the HPHPi optimizer to handle statement
Summary:
The HPHPi optimizer did not allow a statement AST to be replaced by a better
one. The idea is not to disturb the statement so debugger won't be affected.
However, I did find some cases a cheaper statement can be generated without
disturbing the debugger. So I changed the optimizer framework to allow
a statement AST to be replaced (only in the statement list for now). The
constraint still remains though. So we optimize debuggable code rather than
debug optimized code. No real optimization is done in this change.

Test Plan: make fast_tests

Reviewers: mwilliams, qigao

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams, myang

Differential Revision: 316514