[Server] allow 403 being treated as 404
Summary:
As request in task, we want the ability to handle 403 in a way that
people are not sure whether the file exists or not. This diff refactors
the 403 handling code and adds an option to treat 403 as 404.

Test Plan:
make fast_tests
hit the endpoint from the task, show Forbidden when the config is not
set, and show the 404 redirection page when config is set

Reviewers: mwilliams, myang

Reviewed By: mwilliams

CC: ps, mwilliams

Differential Revision: 318674

Task ID: 691676