[Fix] DOMNodeIterator must incRef/decRef the DOM node collection
Summary:
DOMNodeIterator is a wrapper class used to facilitate iteration over DOM
node collections (DOMNodeList or DOMNamedNodeMap). DOMNodeIterator holds a
pointer to the DOM node collection that it is iterating over. The problem
is that the last reference to the DOM node collection can go away while the
DOMNodeIterator is still live, potentially resulting in a use-after-free
which can cause crashes or other bad behavior.

This diff fixes the issue by changing DOMNodeIterator so that it incRefs
and decRefs the DOM node collection to ensure that the collection stays
live.

This diff also fixes a bug where the rewind method wasn't being called when
foreaching over an object that implements IteratorAggregate.

Blame Rev:

Reviewers: myang, mwilliams, kma, oyamauchi, aravind

CC:

Test Plan:
fast_tests
slow_tests
Run www in prod

Revert Plan:

Differential Revision: 322308