Fix infinite recursion in TestCodeError::TestInvalidDerivation()
Summary:
We were using a case-sensitive hash set to store the set of seen base
classes in ClassScope::checkDerivation(). This was causing infiinte
recursion in test cases for invalid derivation such as the following:

class A extends class A {}

This was because the class name was lower case, whereas the parent name
was upper case. Fixed by using a case-insensitive hash map.

Test Plan: make fast_tests

Reviewers: mwilliams, myang

Reviewed By: myang

CC: ps, mwilliams, myang

Differential Revision: 322826

Revert Plan: OK