[Fix] Class hoisting was broken
Summary: We keep a std::set of classes which are hoisted (and therefore dont
need to be checked). The set was case sensitive, but a recent change meant that
it was getting both lower case, and original case names in it. The result was
that we ended up checking the existence of classes which we had already marked
as existing, which was inefficient; and in some cases we checked the existence
of classes before marking them as existing, which was wrong.

Test Plan: fast_tests slow_tests

Reviewers: myang, aravind

Reviewed By: myang

CC: ps, mwilliams, myang

Differential Revision: 324758