[Fix/HPHPi] Fix bug when use $this in closure
Summary:
I was too aggressive when I removed the line that enters the current object
into the assoc list as I thought it is already avaiable in the current env.
This fails when $this appears in the use clause of a closure. Restored the
line. New test added.

Test Plan: make fast_tests

Reviewers: mwilliams

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams, myang

Differential Revision: 324948