[Fix] hphpd crash when throwing non-obj or obj without __toString()
Summary:
Hphpd will hit an exception in the catch block when it handles the case
where an non-object is thrown or an object without __toString method,
causing the DebuggerProxy to disconnect. This diff fixes the problem by
adding another catch.

Test Plan:
make fast_tests
manually run the test in the task and run "throw null"

Reviewers: mwilliams

Reviewed By: mwilliams

CC: ps, mwilliams

Differential Revision: 326091

Task ID: 709819