[Fix] Fix crash in xbox
Summary:
When using xbox for unit test, there was a crash that caused by signal
processing. If the handlers Array is already reset and a signal comes,
in the signal handler we may put another ArrayData * into the Array, which
will no longer be reset before the memory is rolled back. I made a change
to block all signals for this thread so that memory can be safely rolled
back.

Test Plan:
make fast_tests
make slow_tests
run the unit test using xbox that had the crash, no longer see the crash

Reviewers: mwilliams, andrewparoski

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams, andrewparoski, myang

Differential Revision: 327279