Fix TestCodeRunVM-SimpleXML
Summary:
SimpleXMLElementIterator holds a raw C++ pointer to its parent, a
SimpleXMLElement. In expressions like this one:
foreach ($el->children() as $child) { ... }
all references to the SimpleXMLElement returned by children() will be
dropped after the iterator is created and the object will be
destroyed, leaving the iterator with a stale pointer. It looks like
the iterator already had a smart pointer named m_temp that it stuck its
parent in sometimes; this diff gets rid of m_temp and makes sure the
parent is always in a smart pointer. reset_iterator was split into
set_parent and reset_iterator to avoid unnecessary inc/decrefs in
rewind.

Test Plan: TestSimpleXML passes now.

Reviewers: kma, andrewparoski, mwilliams

Reviewed By: andrewparoski

CC: ps, mwilliams, bsimmers, andrewparoski

Differential Revision: 326479

Task ID: 712299