[Perf/cleanup/compile-time] Miscellaneous code reorg
Summary:
Various cleanups
 - switch from __gnu_cxx::hash* to std::tr1::unordered*. The former was 72
bytes, and always allocated at least 53 buckets, while the latter is 48 bytes
and starts with 2 buckets (when asked to start with zero). Since many of these
maps ended up holding a single class, or a couple of functions this was
extremely wasteful of memory. Perf measurements are indicating that its faster
too.
 - change FunctionContainer to hold a StringToFunctionScopePtrMap, rather than
*VecMap, since redeclared functions are rare (non-existent in classes, and
extremely unusual in file-scope). Hold a separate StringToFunctionScopePtrVecMap
(allocated on demand) to deal with the case of FileScope redeclared functions.
 - get rid of .fw.h headers. They are only ever included from one place (the
corresponding .h) and are often empty. This reduces the number of generated
files by about 40%.
 - remove various unneeded forward declarations, and move some virtual
functions out of headers into the .cpp files.

Test Plan: fast_tests slow_tests

Reviewers: qigao, myang

Reviewed By: myang

CC: achao, ps, mwilliams, myang

Differential Revision: 326872