[Perf/HPHPi] Reduce number of user function call lookup
Summary:
When calling a user function in HPHPi through SimpleFunctionCall,
even though we know the name at compile time, we still do a lookup
to find the function statement at runtime. If a function is called
in a loop, the cost of doing lookup becomes more significant. I added
a function table that can be used to reduce lookups. Each user
function name is assigned a unique id (can be reused in case reparsing)
that is used as the index into the function table which is a vector.
Whenever a function declared, we also enter the FunctionStatement into
the function table. In SimpleFunctionCall, we cache the id for the
function being called and use it as a faster way to call the function
if the function is already declared. To prevent a runaway id, I put
a cap as a runtime option. After hitting the cap, no id will be assigned
and those functions have to be called via normal name lookup.

Test Plan:
make fast_tests
make slow_tests
flib/unit test
micro benchmark:
<?php
$start = microtime(true);
function foo() { }
for ($i = 0; $i < 10000000; $i++) foo();
$end = microtime(true);
echo $end - $start . "\n";
Before:
2.4777600765228
After
2.0086898803711

Reviewers: mwilliams, qigao

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams, myang

Differential Revision: 330013