[hphpd] allow accessing base classes' private/protected in bac mode
Summary:
The BypassAccessCheck (bac) mode was added to bypass the check on
private/protected properties and methods so that users can inspect values
and call methods. But it failed to search for properties on base
classes. This diff adds support to do that in the special bac mode and
hide a bunch of warnings.

Test Plan:
make fast_tests
manually verified the case in the task
tried on a small case to access base class's private method

Reviewers: mwilliams, myang

Reviewed By: myang

CC: ps, mwilliams, myang

Differential Revision: 331703

Task ID: 724909