[Fix/HPHPi] Fix bug in FunctionIdTable and VectorConcatExpression
Summary:
This table is request local and cannot use THREAD_LOCAL macros. It needs to
be cleared at the beginning of each request. Otherwise, calling an undeclared
function may not raise an error as it should be if the function was once
declared in previous requests, or it can crash if the previous function was
delete due to reparsing.
The VectorConcatExpression had a bug that it does not check the current m_op
is also a '.', new test case added.

Test Plan:
make fast_tests
flib/unit test
load www sandbox

Reviewers: mwilliams, qigao

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams

Differential Revision: 333027