[Perf] Better class constants
Summary:
If a class constant was defined in terms of something that was not itself a
scalar (eg another class constant), we immediately marked it dynamic, even
though the initializer might eventually be resolved to a scalar.

During optimization, we would only replace a class constant if the value was
scalar, when we can get better results by always replacing.

Type inference allways assumed Variant for dynamic constants, but thats not
necessary. eg "class X { const FOO=1; } class Y { const BAR=X::FOO; }". If X is
volatile, then Y::BAR is dynamic. But it is known to be of type int, and to have
the value 1. Its only dynamic because it might fatal when accessed (if X doesnt
exist).

Any class which contained at least one dynamic constant was marked as
lazy-init, which affected every use of every constant/static in the class.

Test Plan: fast_tests slow_tests

Reviewers: qigao, myang

Reviewed By: qigao

CC: ps, mwilliams, qigao

Differential Revision: 336136