[Fix/Perf/HPHPi] Add ObjectStringPropertyExpression
Summary:
Almost all ObjectPropertyExpression has a explicit string as the property
name. I added ObjectStringPropertyExpression to capture the remaining
cases where the object may not be $this or a simple variable expression
but the property is a string. I also fixed a bug where m_reverseOrder
wasn't set (caused by one of my earlier changes) although I couldn't
find a case where this matters because the use of TempExpressionList
when eval order matters so it may not be needed but leaving it uninited
does not seem right.

Test Plan:
make fast_tests
bench.php showed about 0.5s speed up.

Reviewers: mwilliams, qigao

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams

Differential Revision: 337080