[Fix] Turn off ArgArray
Summary:
The design of ArgArray was flawed because it does not consider the memory
rollback and checkpoint mechanism properly and can leave the memory in
inconsistent state. I am turning it off for now and will reevaluate whether
it is worthwhile to leave it because the removal of collect_few_args_ref
has made it a rare case of packing arguments into an array.

Test Plan:
make fast_tests
run hphpi server mode test several times with no crashes

Reviewers: mwilliams, qigao

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams, myang

Differential Revision: 339617