Add thunk logic to the hphpi binary
Summary:
This diff adds thunk logic so that we can make the hphpi binary redirect
to another binary via execv(). The idea here is that we can set a symlink
"/.hphpi-thunk" on individual machines to make hphpi redirect to a different
binary.

Note that the thunk will only be triggered for the hphpi binary, and only
if "/.hphpi-thunk" exists in the filesystem. Other binaries produced by
hphpc will not trigger the thunk and will continue to behave the same as
before.

Blame Rev:

Reviewers: je, kma, mwilliams, qigao, oyamauchi

CC:

Test Plan:
fast_tests
slow_tests
Verify the thunk is triggered for the hphpi binary
Vierfy the thunk is not triggered for other binaries
Build www and browse the site

Revert Plan:

Differential Revision: 340844