Methods should never be marked as volatile
Summary:
Trying to build phabricator with HPHP results in a compiler error when
compiling "sys/class_map.cpp" saying "class HPHP::GlobalVariables has no
member named fvf_...".

After some investigation, the problem seems to be that methods are getting
marked as volatile (because AllVolatile=true) and the logic for emitting
g_class_map tries to reference non-existent fvf_ variables. I was able to
write a small program that repros the issue when AllVolatile=true:

  <?php
  class C { private function foo() {} }
  $obj = new C;
  var_dump($obj);

As best as I can tell, HPHP's plumbing doesn't really support marking
methods as volatile, and furthermore the concept of marking methods as
volatile seems unnecessary. It is not possible to conditionally define a
method, and a method is considered defined iff its enclosing class is
defined, so it seems like the logic for dealing with volatile classes and
redeclared classes will take care of methods for us.

This diff fixes the issue by updating the compiler so that methods are
never marked as volatile.

Blame Rev:

Reviewers: mwilliams, myang, kma

CC:

Test Plan:
make slow_tests
Phabricator build succeeds

Revert Plan:

Differential Revision: 341219