[Fix] Allow private static properties to be found by dynamic lookup
Summary:
os_lval and os_get have always excluded private properties, presumably because
they would normally be resolved statically. There are cases however (eg
involving redeclared classes, or $cls::$property) where we cant resolve it
statically, and so we should allow privates to be found.

This fix isnt exactly right, since it could find a public property in a more
derived class, when it should have found a private property with the same name
(it also allows access to a private property that should be hidden). But this
fixes the immediate issue.

Test Plan: fast_tests slow_tests

Reviewers: kma, qigao, myang

Reviewed By: myang

CC: ps, mwilliams, myang

Differential Revision: 342147