[Fix] Only optimize dynamicNamedCall when its known to be dynamic
Summary: There was a small optimization for the case of a dynamic call through
a derived-from-redeclared class. But its only valid when the call is known to be
dynamic. The objCall flag is used to indicate dynamic (>0) possibly-dynamic
(<0), and static (0), but we were just testing for non-zero. Replace with a test
for >0 instead.

Test Plan: fast_tests slow_tests

Reviewers: myang, kma

Reviewed By: myang

CC: ps, mwilliams, myang

Differential Revision: 342326