[Fix] Do not set HTTP_RAW_POST_DATA if it cause String overflow
Summary:
In video upload, the amount of data can overflow a String and cause exception
to be throw. I made a change so that we don't store HTTP_RAW_POST_DATA when
this happens.

Test Plan:
make fast_tests
upload a large video in non-chunk mode and make sure no exception is thrown
at the end.

Reviewers: mwilliams, qigao

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams

Differential Revision: 345263