[Fix/hphpd] remove the trailing ';' for '=' command
Summary:
In a recent change, () were added to wrap the expression for '=', which
made ';' in the expression a syntax error. We were thinking on a general
fix but it seems hard to do without calling php parser for the command
line. This diff fixes one common benign case with traling ';', which used
to work before the () were added.

Test Plan:
make fast_tests
tried with hphpd on command line
hphpd> =123;
123

Reviewers: mwilliams, myang

Reviewed By: myang

CC: ps, mwilliams, myang, qigao

Differential Revision: 351726

Task ID: 747161