[APC] Improve apc size profiling
Summary:
This diff does a few things to improve apc size profiling:
1. More compact format for script to process the result. I changed it to
one entry a line and got rid of the global object in json.
2. Add a profiling to monitor the size of keys with noTTL in a group
3. Switch to use concurrent_hash_map to protect the group and detail
profiling
4. change some field from int64 to int32 to save some memory

Test Plan:
make fast_tests
production testing on profiling
Note that all the code touched in this diff is not on critical path if
profiling is turned off. This is staging for enabling profiling on h2g
machines.

Reviewers: myang, mwilliams, ps

Reviewed By: mwilliams

CC: ps, mwilliams

Differential Revision: 352855