Fix for code generation when PregenerateCPP=true.
After introduction of property tables the current way of c++ files generation
when clustering is requested together with PregenerateCPP=true does not work.

It happens because after code is pregenerated the classes scope information
gets lost, i.e. not available for code generator when it builds property tables.

The fix saves the classes scope when pregeneration is done and restore it back
to code generator when that pregenerated code used, making it possible to build
property tables after.