Bug in accessing constant from eval statements.
The following code
<?php
   define('A', 'test');
   eval("var_dump(A);");
?>
will result in
HipHop Notice:  Use of undefined constant A - assumed 'A' in string on line 1
string(1) "A"
when compiled with eval enabled.

The bug was introduced with 1a09996cba4dea52c9765997f006e8228d417a43.

The fix provide gives it another try to find constant in dynamic table
before issuing an error.