Bug in handling non-static call as a static in some edge cases.
The following example will result in a segmenation fault.

<?php
   class A {
      public function f($x) {
         return $x;
      }
   }
   class B {
      function __construct() {
         array_map(array('A', 'f'), array(0,1,3));
      }
   }
   $b = new B();
?>

The problem seems to be in the way array_map works. It creates and initialize
cmp for callback once for all calls. Then during the first call the dummy
object is created and the reference to corresponding ObjectData is saved within
mcp. This dummy object is destroyed right after the call is done, but the mcp
still has that reference, which is used for next calls, which causes problem.