ScopeIdPrefix option and mangle dollar sign
ScopeIdPrefix added to CodeGeneration section,
which is used in getId(), getDocName() and
getInjectionId() methods in scope classes.
It sets a prefix for m_redeclaring counter in
ids.

Option::MangleFilename changed to replace dollar and
and slash with underscore in OSS version, becuase
dollar sign in filename must be escaped but cmake
generates files with unescaped filenames. (depend.make)