Use "guarded" information for OpThis and RetC
Static analysis computes a "guarded" flag for various constructs
that can fatal (such as $this->) to indicate that such a construct
has already been seen.

We can use this flag in the translation of OpThis to avoid repeatedly
testing for null, and in the translation of OpRetC to unconditionally
decRef ar->m_this.

Also fixed some compiler generated byte code for closures that was
emitting strings for vector operations (rather than using literals
in the vectors).