Separate dwarf infor for a and astubs
gdb tends to get confused if we emit dwarf info
for a and astubs together in the same ELF file.
gdb assumes that all ELF files emitted contain
code addresses in contiguous range.
If a single ELF file contains dwarf file for both a and
astubs, gdb assumes all intermediate addresses must
also belong to this ELF file, and this causes
symbol name lookup in gdb to fail for intermediate
addresses that may be present in other ELF files.