Streamline copyImpl() code in HphpArray
This is a code cleanup patch to make it easier to spot
bugs and refactor code.

Numerous functions in HphpArray check to see if they must
copy the array first, then contain two whole copies of
the function body, one if we copied and one if we didn't.

There's already a coding style in some functions that avoids
the duplicated code;  this patch uses that style in the rest
of the code.

Early results show this is about 0.1% increase in CPU
Instructions but about 1% decrease in CPU Time, although
some endpoints have 1% noise