Fix a bug in nested VM exception handling
It was possible for C++ exceptions to propagate when setting
up a nested VM, but the nested VM didn't handle it and just let it fly
out to the containing VM.  The previously nested VM would end unwinding
after this first nested ActRec, and then popVMState as if it were the
nested VM.  The net result is we would unwind for all but the first VM
entry, because the m_nestedVMs state was off by one.

This fixes a test that crashed the JIT:  in that case, we were
re-entering to try to generate the 500 page, attaching the global VarEnv
to some random ActRec that should've been unwound, and then throwing again
(so this was noticed by seeing the VarEnv depth refcounts going wrong).