Inline small HphpArrays
HphpArray uses a minimum table size of 8 elements and always allocates using
malloc.  Instrumentation shows that 85% of arrays are <= 4 elements,
and if we make room for a 4-entry table (up to 3 elements) inline,
we can save a bunch of cycles and memory for small arrays.

Initially we use the inline table.  We also union the inline table with
a medium sized hashtable, but without slots, so for medium sized arrays,
we only need to malloc memory for the slots, not the hashtable.

This change also reduces MinLgTableSize from 3 to 2, to reflect the
smaller initial sizes we support, and fixes a small memory accounting
but where HphpArray::sweep() didn't call adjustUsageStats().