Inline-allocate up to 4 Bucket pointers in ZendArray
HphpArray and VectorArray benefit from inline-allocating
so lets do it in ZendArray.  This inline-allocates just
the bucket pointer array, up to four elements.

It also drops the m_nTableSize field since it's always
equal to m_nTableMask + 1, and shuffles fields in ZendArray
to put hot fields near the beginning of the object.