Fix COW bug with foreach by reference
There are some cases involving foreach by reference loops where COW is not
triggered when it should be. Here is an example:

  <?php
  function foo() {
    $arr = array(0,10,20,30);
    foreach ($arr as $k => &$v) {
      $v += 100;
      if ($k == 1 && !isset($arr2)) { $arr2 = $arr; }
    }
    unset($v);
    var_dump($arr2);
  }
  foo();

  === Current output under HipHop ===
  array(4) {
    [0]=>
    int(100)
    [1]=>
    int(110)
    [2]=>
    int(120)
    [3]=>
    int(130)
  }

In the example above, the 3rd and 4th elements of $arr2 were incorrectly
modified by the foreach by reference loop. The problem is that the strong
iterator only checks if the array's refcount is greater than 1 before the
first iteration of the loop, but it doesn't check the array's refcount for
subsequent iterations.

I fixed this by making the strong iterator check the array's refcount each
time the iterator is advanced. If the refcount is greater than 1, a copy is
made and the strong iterators are migrated from the original array to the
copy.

In the course of implementing this fix, I made VectorArray escalate to
ZendArray whenever it is used with foreach by reference. This helped avoid
complications in cases where a VectorArray would need to escalate to
ZendArray while a foreach by reference loop is iterating over it. I also
took the opportunity to clean up some of the logic that deals with
escalating an array when a foreach by reference loop begins.