Use shorter encoding for "mov imm32,reg32", and improve jmp() and call()
During a recent experiment I copied disassembly of some code generated by
HHVM into a .S file and compiled it, and I noticed that we were using a
6-byte encoding for "mov imm32,reg32" when a 5-byte encoding was possible.
This diff fixes the x64 assembler to take advantage of the better encoding
for "mov imm32,reg32".

I also noticed that the jmp() and call() methods in X64Assembler don't take
advantage of emitImmReg goodness, so I updated them appropriately.