Dont crash on incorrect type assertions
Given something like:

  if ($this instanceof foo) {
  }

We can infer the type of this inside the if block to be foo, assuming
that foo is stronger than the known type of $this.

But if foo is a completely unrelated class, we shouldnt do that, because
we could generate incorrect code at compile time (of course, we could also
optimize out the if entirely - but currently dont).

The code to reconcile an actual type with an asserted type assumed the
actual type was passed first, but in one place, we passed the asserted
type first. This was only a problem if the assertion was impossible.