Fix a segfault in get_declared_classes()
In rare circumstances, it could attempt to read an invalid iterator.
The problematic case is where a function or class has a name that
puts it at the "end" of the NamedEntityTable (the order of which is
undefined to start with) - and if its a class, its definition must
have been seen, but it must not have been declared yet.