The generated 86ctor method should not be callable
The 86ctor method is only there for the benefit of FPushCtor*, it
should not be called via parent::__construct.

This removes it from the method table altogether, since we never
need to look it up dynamically; which also fixes:

  $ctor = '86ctor';
  $x->$ctor();

While I was there, I also removed 86pinit and 86sinit. For now, Im
leaving 86cinit, since it would mean increasing the size of Class
(by one pointer).

This also revealed a bug - 86ctor was not supposed to be inherited,
but it could be if the emitter "forgot" to create one when needed -
or if the user failed to create one using hhas. One of the hhas
examples violated this. I changed the base class constructor to be
__construct instead of 86ctor.