Add support for imageantialias
This was disabled because it requires a special php version
of the gd library. But the existing gd library has similar
functionality via a different api. Since there are only
a few functions that are affected by antialiasing it was
fairly easy to intercept each of the affected routines and
do the right thing.