Allocate ThreadLocalSingleton storage inline
ThreadLocalSingleton stores a __thread T* and lazy-initializes the
object by calling T::Create, which typically allocates a new object.
Access to the thread local T is indirect through this pointer.

We can do better by just inline-allocating __thread T.  This patch
does so while preserving the timing of when T is initialized or
destructed.