Annotate more variables as UninitNull.
I noticed that we weren't annotating the type of
an FPassL even if the variable was known to be UninitNull.

Fixing that exposed a bug in the metadata argument numbering,
and existing tests in TestCodeRun revealed another bug in type
inference, in scopes where $this is altered via a dynamic
variable ($t='this'; $$t = 'surprise').