Fix use-after-free bug in assignRefHelper
assignRefHelper() takes care to increment the refcount before
copying a value in case destroying the LHS destroys the owner
of the ref; but then it reads from said owner after destruction,
which is a free-after-use bug, for example:

  $x = array(0);
  $x &= $x[0];

It never showed up before because SmartAllocatorImpl::dealloc().
did not garbage-fill freed memory.  We just need to save the
newly created pointer before destroying the LHS value.

This patch also enables garabge-filling for
SmartAllocatorImpl::dealloc().