De-unionize StringData.m_hash and m_shared
Now that we have inline-string space in StringData, we can use it
to store the SharedVariant* m_shared field for shared strings,
shrink m_hash to 32 bits, shuffle fields around so they still
pack nicely, then increase the inline space in StringData from
40 bytes to 44 bytes.

Doing this requires reducing string hashes to 32 bits everywhere,
which means 31 bits in practice because we use the high bit as a
flag in various places.  (as before... we're really going from
63 bits to 31 bits).

The change also ripples into HPHPC tables which contain a string
hashcode since we can shrink that field to 32 bits.