Protect the apc expiration priority queue
It's possible to store the same key over and over again with a
ttl and it will grow the expiration queue unbounded.  Use a concurrent
hashtable to store items which are already in the queue to prevent
adding the same key over and over.  We only honor the first TTL and
ignore the rest.