Store Continuation locals inline at the end of the object
The call to free() in ~c_GenericContinuation was a nontrivial
source of dTLB misses. The locals previously stored in that malloced
buffer are now stored in extra space allocated at the end of the
object, just like builtin properties at the end of an Instance. This
also makes translated code touching these locals slightly faster by
replacing a mov with an lea.