Lock all targetcache allocations
There were a few places (including Cache::alloc and
allocStatic) that allocated space in the targetcache
without taking the handleLock. This was mostly ok,
because they were only called when we held the
write lease; but there are paths in the interpreter,
the file cache and Unit::merge that can allocate
target cache entries without holding the write lease.

This resulted in hard to track inconsistencies in
targetcache state.