Handle inheriting from GenericContinuation correctly.
GenericContinuation does custom deallocation of memory, but did
so assuming it would not be inherited from. While inheriting is useless for
end-users, we implement some real continuations by inheriting from
GenericContinuation, so simply making it final is not an option.

(This appeared to work before because we weren't testing with enough fields
in the child class.)