Parser bug fixes and clean up
I combed through HipHop's grammar and noticed a few bugs. HipHop's grammar
does not allow for traits to be declared anywhere except toplevel scope
(whereas Zend allows for this), and the grammar allows for bogus singleton
XHP tags such as '<foo a="1" /foo>' (which are not allowed by the original
XHP grammar given by xhpast).

This diff fixes these bugs, and it also removes some superfluous rules,
duplicated rules, and dead rules from the grammar.

I also took a look at 14 shift/reduce conflicts in the grammar to see if I
could eliminate most of them. 4 of the 14 conflicts went away when I fixed
the issue with the xhp_tag_body rules and when I removed the dead type_decl
rules. Of the 10 remaining conflicts, 2 were due to the if/else rules and
the other 8 were due to the "expr -> expr [ dim_offset ]" rule. Conflicts
due to "if/else" are well understood and conventional wisdom says we're
better off just accepting these conflicts instead of contorting the grammar
to try to get rid of them. Conflicts due to "expr -> expr [ dim_offset ]",
however, are not really well understood and there are a lot of them, so it
seemed worth fixing.

"expr -> expr [ dim_offset ]" was originally added to the grammar to
support using function calls as the base for array element expressions
(ex. "$x = f()[0]"). The rules for variable, calls, array elements, and
object properties were written in a very weird way that made them hard to
update, so I took the opportunity to rewrite these rules in a more natural
form and I updated the rules to support using function calls as the base
for array element expressions. Finally, I ran into cases where developers
were using array literals, class constants, and parenthesized expressions
as the base of array element expressions, so I introduced 'dim_expr' rules
so that HipHop's grammar will continue to support these cases.