Make ArrayData::m_strongIterators a linked list instead of vector
For the cost of one more pointer in struct FullPos, we can simplify
the tracking of an array's currently active mutable iterators.
The current array (vector of FullPos*) requires a secondary malloc()
if there are >1 iterators.  Using a linked list never requires malloc,
and keeps the common case fast (most recently added iterator is the first
one in the list).