Added isContextFixedCheck to analyze[CGetS, SetS] and cleaned up uses of arGetContextClass.
The static property accessor bytecode translations were not
checking whether the context is fixed, which is incorrect when translating
pseudomains. This diff adds a isContextFixed check to analyze[CGetS, SetS] to
make sure we are not in pseudo-main. It also replaces uses of arGetContextClass
with curFunc()->cls() where its clear that we don't want to be in pseudomain.