Disallow using comparison operators on collections
Using comparison operators (==, !=, <>, <, <=, >, >=) with arrays and
objects can be expensive and produce strange results in some cases.

Rather than having collections match this behavior, I think it's better to
disallow using these operators to compare a collection with an integer,
string. This diff adds checks to either raise a warning or throw an
exception (depending on a runtime option) when any of these operators is
used to compare collection with an integer, string, array, or object.

Note that collections are still allowed to be compared with null, true, and
false using the comparison operators. Also, collections can still be
compared for reference equality with === and !==.