Dont call getMInstrCtx unless necessary
The translations for the M* instructions call getMInstrCtx to
set MInstrState::ctx if the context is not fixed, or if there
are property accesses in the vector. But MInstrState::ctx is
only read when the context is not fixed AND there are property
accesses in the vector.

Also found an issue where the BaseN* helpers would access
the current frame without a VMRegAnchor. Fixed by passing in
the correct fp as a parameter.