Add HPHP::safe_cast and use it in the assembler
This is a wrapper around boost::numeric_cast, to give a
helpful error message if we attempt to use an out of range immediate
or jump in the assembler. I'm sure there are other places in the
codebase it can be useful but this was the first that came to mind. In
RELEASE=1 builds it's just a static_cast so there should be no runtime
overhead.