Only validate that the subject of preg_split is valid utf8 once per call
pcre_exec checks that its subject is valid utf8 every time
it's called (unless told not to). Our preg_split implementation can
end up calling pcre_exec once for every character in the subject for
certain inputs, so rechecking the utf8 validity of the subject is a
waste of cycles.