Add syntax for directly calling a closure stored in a property
When a closure is stored in a property (ex. "$o->p = function(){..};"),
there is no way to call the closure directly. Currently, if the user wants
to invoke a closure stored in a property, an intermediate local variable
must be used like so: "$cl = $o->p; $cl();".

This diff tweaks HipHop's grammar to be a little more flexible so that it
is possible to directly invoke a closure stored in an object property. If
a closure stored inside "$o->p", the closure can be invoked like so:
"($o->p)();". The parentheses wrapped around "$o->p" lets the parser know
that we are not calling a method named "p", but rather we are invoking a
closure stored in a property named "p".

This diff also introduces similar syntax for invoking a closure stored in
a static property ("(C::$x)()") or in an array inside a static property
("(C::$x[0])()").