Fix an issue with over-agressive type annotations
This could cause a local with a type assertion to be
inferred non-null at places where thats not the case.

In the new test, the $a in "return $a" was inferred to
to be an object, resulting in a crash when $a was
pushed onto the stack, and its refcount incremented.