Fix bug in DebuggerClient's use of Sweepable Socket objects
DebuggerClient attempted to avoid problems with destructing
Socket objects by calling incPersistent so they would not be swept.
However, the semantics of incPersistent apparently require that the
object only is persistent in some sort of thread-local cache (it is
re-added to the sweepable list for the same thread immediately).  When
we disconnect the debugger, the DebuggerClient is deleted and the
socket object is also destructed---the Sweepable destructor tries to
remove it from the current thread's list, so if this is a different
thread, the next sweep loop on the original creating thread will make
function calls on a deleted object.