Fix injected bug: check m_strongIterators on the copy
This was injected when I converted the strong iterator
list to a linked list.  Only affects HphpArray when
COW occurs on the array being mutated in a mutable foreach
loop.  Starting a strong iteration causes a private copy
to be made anyway, but the bug should be fixed either way.