Use libcurl 7.28.0's curl_multi_wait() or fallback when curl_multi_select not available
Removes the need to patch libcurl and build a copy
explicitly for HipHop.

If libcurl >= 7.28.0 we get the new curl_multi_wait() function.
If not, we either get the patched in curl_multi_select(), or a
fallback which works for FDs < 1024, but fails (with an error
message) for sites which use more concurrent FDs.