Fix crash involving static properties with traits
Some programs that used static properties with traits would cause the VM
to crash while building the Class. Here is an example:

  <?php
  final class Foo {
    use Bar;
    private static $a = array();
  }
  trait Bar {
    private static $a = array();
  }

The crash was happening because was the m_staticProperties map of the
current class was being accessed before it was initialized. This fixes the
logic to directly read m_val from the SProp structure for static properties
declared by the current class to avoid this problem.