Make call_user_func_array less strict
If the called function expects a reference,
but doesnt get one, zend raises a warning,
and returns false, rather than executing the
body of the function.

A previous diff emulated that behavior, but we have
a lot of tests that rely on the function being
called anyway. This version still raises the
warning (so we'll hopefully get the bugs fixed)
but continues on and executes the function.