Fix various issues with the call_user_func optimization
For a call_user_func to a simple function which
didnt exist at the time of the call, we gave the
wrong info to recordReentrantStubCall, resulting
in a corrupt stack.

fb_call_user_func_safe_return didnt always return the
correct result in jitted code if the function existed
when the translation was generated, but doesnt exist
at a later call, and the default value was non-null.

fb_rename_function simply didnt work on any jitted
*call_user_func* code.

The warning for call_user_func* given an unknown function
was different between jitted and interpreted code, resulting
in verify_quick failures.