Move some of our ABI constants to an abi-x64 header, and random code cleanup
This was extracted from my working branch on returning values
in registers, without anything related to that part of it.  It is just
a bit of code cleanup and allows sharing a few of the register
constants with the ir, and a little bit of additional documentation on
a few things.  (There are probably more things that should still be
pulled to shared locations between the two translators, and I'm sure
there's still a fair bit of distributed knowledge about which
registers are what.)