Choose helper reffiness with types.
Some elem/prop helpers know that their keys must be cells, but
those that use locals could get passed Refs. We know the reffiness of
the input keys at translation-time; use this type, rather than the input
flavor, to choose which helper to call. This should strictly reduce the
number of KindOfRef checks in vector instructions.

To gain confidence this was working as intended, I also factored out the
ref checks into a helper that asserts that it has stripped any refs.