Increase StringData size limit to 2GB (2^31-2).
This diff simply increases the string size limit, and cleans
up a handful of places we used int when we should have used
uint32_t anyway. This clears the way for a followup, which makes
limits StringBuffer to the same length limit as StringData.

We use 2^31-2 because the real limit is 2^31-1 (biggest unsigned
value for a signed-int length), but we want to safeguard against
the likelihood of client code computing size()+1 into a signed-int32
and wrapping.