Inline some int-keyed HphpArray accesses.
Stay in translator output for some simple array accesses. For
now, this just further specializes the already-specialized case of
array_getm_i. We'll build on this in subsequent diffs.

The translation inlines the *entire* array access, quadratic probe and
all. This appears necessary; only doing one probe, and then falling back
to a helper did not seem to win.