Support statically-known Class* information in more M-vector Prop operations
Get more information about statically known object types from
hphpc to the JIT.  Uses this information on intermediate prop
operations for all M-vectors, and final operations for CGetM and SetM.
Removing the old SetMProp specialization now is a win.  I also tried
implementations of the final operations for IssetEmptyProp, VGetProp,
BindProp, and UnsetProp, but none of them seemed to pay (and VGetProp
didn't work) so I'm leaving them out of this.