Pass HPHPc's function return types to HHVM
HPHPc can often infer function return types, although those types might
not be 100% accurate because values can also be Null.  This diff stores
those types in the bytecode meta-data and uses them in the JIT to
generate better code.

In the general case where values may also be Null, the information is
passes as a prediction.  In some cases, HPHPc guarantees that the type
is Null, so these cases are passes in as inferred and not predicted to
save the check altogether.

Also, if HPHPc's predicted type is not ref-counted, we know for sure that
the values is not ref-counted (because it's either that type or Null).
So use that fact to avoid generating any code for PopR in those cases.