Eliminate Jcc to Jmp for conditional branches at the end of a trace.
In eliminateDeadCode(), identify conditional branches at trace ends
that have a side-effect free trace exit.  Connect the trace exits to the
Jcc instruction by appending its result SSATmp* to the exit's list
of inputs and set the value to IR_DIRECT_JCC_JMP_ACTIVE.

Option -- While the SSATmp linkage works and makes it easy to visualize
with Eval.DumpIR=1, I'm considering removing it. Connecting the instrs
using the m_tca in each instruction is simpler.
In addition, linking via the result SSATmp interferes with converting
guard checks to direct branches since these operations are currently
combined with the loads.  May be better to split these into a guard
check with a label followed by a separate load.

Combine cgJmpZero and cgJmpNZero into cgJmpZeroHelper(.., cc);
Use LabelInstruction::m_patchAddr rather than m_asmAddr to anchor
the list of addresses that need to be patched to point to the label.