Dont assume the class exists in FPushClsMethodD
If the class exists at translation time, and is marked
AttrUnique we would burn in the address and assume its
always defined - but it may not be. So add the check.

hphpc does some static analysis to determine that the
class really is available, so use that to skip the check
when possible.

In addition, do some cleanup on Self, Parent, AGetC
and FPushClsMethodF.