Update outdated .inc files and generated files
Some of the .inc files were out of date after KindOfInt32 was removed. Some
C++ functions use "int" for one or more parameters (we left in support for
Int32 parameters), and this diff updates the .inc files to reflect that.

This diff also updates a few .inc files that were out of date after the
__destruct method was removed.