Allocate more bits at a time.
We've seen a lot of LLC misses touching the bitmaps in
the target cache. I'm working on a more elaborate change
to type-segregate the targetcache, but meanwhile this
simple diff seems to capture some of the opportunity.

We know we'll be allocating a lot of these bits,
so allocate them in great, throbbing chunks.