Fix a race condition
TypeConstraint's constructor is not thread-safe on the
first call. This was causing infrequent hangs for repo-
authoritative bulids.

There may be a similar issue in non-repo-authoritative
runs - but Ive never hit it. I suspect that loading
systemlib.php prevents issues. But Ive added an explicit
beginning of time initialization anyway.