Fix use of funcd as a flag for FPushClsMethodF
I used funcd on an FPushClsMethodF as a flag to indicate
that the destination of the call was known at compile time.

But if the FPushClsMethodF is in the fpi region of another
call that was fully bound, funcd will be set on it anyway.

FPushClsMethodF is almost always directly bound, so this
wasnt causing too much trouble. Pretty much the only time
its not directly bound is when the method doesnt exist
(including the case where it doesnt exist, but is handled
by __call or __callstatic).