Fix throwing from FunctionEnter/Exit
In a really hard to reproduce edge case, fcallHelper would call
doFcall, which could then throw from FunctionEnter via
CheckSurprise.

In this case, the stack wasnt setup correctly to unwind
through fcallHelper, and the c++ exception mechanism called
std::terminate.

To fix it, I added a try catch to fcallHelper, and smashed
the return address to be the call's true return address.
This allows our unwinder to be called, and everything's happy.

While there, realized that we were swallowing exceptions from
the user profiler (basically to avoid issues like this); but
doing so was pointless, given that CheckSurprise could still
throw.