Avoid using MInstrState for statically known CGetMs
Many vector instructions don't need an MInstrState struct on
the stack to be executed. The most straightforward of these are CGetMs
with all of their offsets known at compile time, which is what this
diff adds support for. Now the generic vector translator is fast
enough that we can permanently kill off CGetMProp (I've manually
verified that the assembly output for the fast case is exactly the
same). isSupportedCGetMProp is still around for the IR.