Fixed bugs in property accesses, conversion and simplification, and added additional runtime asserts to generated code.
Fixed a bug in static property accesses where we passed the
wrong parameter to SPropCache::lookup. Fixed some code gen problems in
type conversions. Added a check stack that checks top of stack contents
for bad ref counted values. This is called on a function return.
Fixed a refcounting bug in int-to-string casting and CSE'ing of instructions
that produce or consume reference count.