String equality can be improved.
Ben Maurer pointed out an opportunity this morning while looking at perf top
output: eq_str_str was a little too quick to resort to memcmp. It was relying
on the trinary StringData::compare(), which falls back to
memcmp(a, b, min(len1, len2)) once the easy cases have been exhausted.

If we don't care about which direction the inequality goes (as for ==, !=,
etc.), we can first test length. We apparently test unequal strings frequently
enough for this to be profitable. I'll also try comparing hashes.

This is a little rough as-is; for instance, the short-circuit test for
identical pointers in eq_str_str might not really be helping.